Add or replace data frame columns

add_columns() combines two or more data frames, but unlike cbind or bind_cols, this function binds data as last columns of a data frame (i.e., behind columns specified in ...). This can be useful in a "pipe"-workflow, where a data frame returned by a previous function should be appended at the end of another data frame that is processed in add_colums().

replace_columns() replaces all columns in data with identically named columns in ..., and adds remaining (non-duplicated) columns from ... to data.

add_columns(data, ..., replace = TRUE)

replace_columns(data, ..., add.unique = TRUE)

Arguments

data
A data frame. For add_columns(), will be bound after data frames specified in .... For replace_columns(), duplicated columns in data will be replaced by columns in ....

...
More data frames to combine, resp. more data frames with columns that should replace columns in data.

replace
Logical, if TRUE (default), columns in ... with identical names in data will replace the columns in data. The order of columns after replacing is preserved.

add.unique
Logical, if TRUE (default), remaining columns in ... that did not replace any column in data, are appended as new columns to data.

Value

For add_columns(), a data frame, where columns of data are appended after columns of ....

For replace_columns(), a data frame where columns in data will be replaced by identically named columns in ..., and remaining columns from ... will be appended to data (if add.unique = TRUE).

Note

For add_columns(), by default, columns in data with identical names like columns in one of the data frames in ... will be dropped (i.e. variables with identical names in ... will replace existing variables in data). Use replace = FALSE to keep all columns. Identical column names will then be renamed, to ensure unique column names (which happens by default when using bind_cols). When replacing columns, replaced columns are not added to the end of the data frame. Rather, the original order of columns will be preserved.

Examples

data(efc)
d1 <- efc[, 1:3]
d2 <- efc[, 4:6]

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

head(bind_cols(d1, d2))

#>   c12hour e15relat e16sex e17age e42dep c82cop1
#> 1      16        2      2     83      3       3
#> 2     148        2      2     88      3       3
#> 3      70        1      2     82      3       2
#> 4     168        1      2     67      4       4
#> 5     168        2      2     84      4       3
#> 6      16        2      2     85      4       2

add_columns(d1, d2)

#> # A tibble: 908 x 6
#>    e17age e42dep c82cop1 c12hour e15relat e16sex
#>     <dbl>  <dbl>   <dbl>   <dbl>    <dbl>  <dbl>
#>  1     83      3       3      16        2      2
#>  2     88      3       3     148        2      2
#>  3     82      3       2      70        1      2
#>  4     67      4       4     168        1      2
#>  5     84      4       3     168        2      2
#>  6     85      4       2      16        2      2
#>  7     74      4       4     161        1      1
#>  8     87      4       3     110        4      2
#>  9     79      4       3      28        2      2
#> 10     83      4       3      40        2      2
#> # ... with 898 more rows

d1 <- efc[, 1:3]
d2 <- efc[, 2:6]

add_columns(d1, d2, replace = TRUE)

#> # A tibble: 908 x 6
#>    c12hour e15relat e17age e42dep c82cop1 e16sex
#>      <dbl>    <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1      16        2     83      3       3      2
#>  2     148        2     88      3       3      2
#>  3      70        1     82      3       2      2
#>  4     168        1     67      4       4      2
#>  5     168        2     84      4       3      2
#>  6      16        2     85      4       2      2
#>  7     161        1     74      4       4      1
#>  8     110        4     87      4       3      2
#>  9      28        2     79      4       3      2
#> 10      40        2     83      4       3      2
#> # ... with 898 more rows

add_columns(d1, d2, replace = FALSE)

#> # A tibble: 908 x 8
#>    e15relat e16sex e17age e42dep c82cop1 c12hour e15relat1 e16sex1
#>       <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>     <dbl>   <dbl>
#>  1        2      2     83      3       3      16         2       2
#>  2        2      2     88      3       3     148         2       2
#>  3        1      2     82      3       2      70         1       2
#>  4        1      2     67      4       4     168         1       2
#>  5        2      2     84      4       3     168         2       2
#>  6        2      2     85      4       2      16         2       2
#>  7        1      1     74      4       4     161         1       1
#>  8        4      2     87      4       3     110         4       2
#>  9        2      2     79      4       3      28         2       2
#> 10        2      2     83      4       3      40         2       2
#> # ... with 898 more rows

# use case: we take the original data frame, select specific
# variables and do some transformations or recodings
# (standardization in this example) and add the new, transformed
# variables *to the end* of the original data frame
efc %>%
  select(e17age, c160age) %>%
  std() %>%
  add_columns(efc)

#> # A tibble: 908 x 28
#>    c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4
#>      <dbl>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1      16        2      2     83      3       3       2       2       2
#>  2     148        2      2     88      3       3       3       3       3
#>  3      70        1      2     82      3       2       2       1       4
#>  4     168        1      2     67      4       4       1       3       1
#>  5     168        2      2     84      4       3       2       1       2
#>  6      16        2      2     85      4       2       2       3       3
#>  7     161        1      1     74      4       4       2       4       1
#>  8     110        4      2     87      4       3       2       2       1
#>  9      28        2      2     79      4       3       2       3       2
#> 10      40        2      2     83      4       3       2       1       2
#> # ... with 898 more rows, and 19 more variables: c86cop5 <dbl>, c87cop6 <dbl>,
#> #   c88cop7 <dbl>, c89cop8 <dbl>, c90cop9 <dbl>, c160age <dbl>, c161sex <dbl>,
#> #   c172code <dbl>, c175empl <dbl>, barthtot <dbl>, neg_c_7 <dbl>,
#> #   pos_v_4 <dbl>, quol_5 <dbl>, resttotn <dbl>, tot_sc_e <dbl>, n4pstu <dbl>,
#> #   nur_pst <dbl>, e17age_z <dbl>, c160age_z <dbl>

# new variables with same name will overwrite old variables
# in "efc". order of columns is not changed.
efc %>%
  select(e16sex, e42dep) %>%
  to_factor() %>%
  add_columns(efc)

#> # A tibble: 908 x 26
#>    c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4
#>      <dbl>    <dbl> <fct>   <dbl> <fct>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1      16        2 2          83 3            3       2       2       2
#>  2     148        2 2          88 3            3       3       3       3
#>  3      70        1 2          82 3            2       2       1       4
#>  4     168        1 2          67 4            4       1       3       1
#>  5     168        2 2          84 4            3       2       1       2
#>  6      16        2 2          85 4            2       2       3       3
#>  7     161        1 1          74 4            4       2       4       1
#>  8     110        4 2          87 4            3       2       2       1
#>  9      28        2 2          79 4            3       2       3       2
#> 10      40        2 2          83 4            3       2       1       2
#> # ... with 898 more rows, and 17 more variables: c86cop5 <dbl>, c87cop6 <dbl>,
#> #   c88cop7 <dbl>, c89cop8 <dbl>, c90cop9 <dbl>, c160age <dbl>, c161sex <dbl>,
#> #   c172code <dbl>, c175empl <dbl>, barthtot <dbl>, neg_c_7 <dbl>,
#> #   pos_v_4 <dbl>, quol_5 <dbl>, resttotn <dbl>, tot_sc_e <dbl>, n4pstu <dbl>,
#> #   nur_pst <dbl>

# keep both old and new variables, automatically
# rename variables with identical name
efc %>%
  select(e16sex, e42dep) %>%
  to_factor() %>%
  add_columns(efc, replace = FALSE)

#> # A tibble: 908 x 28
#>    c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4
#>      <dbl>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1      16        2      2     83      3       3       2       2       2
#>  2     148        2      2     88      3       3       3       3       3
#>  3      70        1      2     82      3       2       2       1       4
#>  4     168        1      2     67      4       4       1       3       1
#>  5     168        2      2     84      4       3       2       1       2
#>  6      16        2      2     85      4       2       2       3       3
#>  7     161        1      1     74      4       4       2       4       1
#>  8     110        4      2     87      4       3       2       2       1
#>  9      28        2      2     79      4       3       2       3       2
#> 10      40        2      2     83      4       3       2       1       2
#> # ... with 898 more rows, and 19 more variables: c86cop5 <dbl>, c87cop6 <dbl>,
#> #   c88cop7 <dbl>, c89cop8 <dbl>, c90cop9 <dbl>, c160age <dbl>, c161sex <dbl>,
#> #   c172code <dbl>, c175empl <dbl>, barthtot <dbl>, neg_c_7 <dbl>,
#> #   pos_v_4 <dbl>, quol_5 <dbl>, resttotn <dbl>, tot_sc_e <dbl>, n4pstu <dbl>,
#> #   nur_pst <dbl>, e16sex1 <fct>, e42dep1 <fct>

# create sample data frames
d1 <- efc[, 1:10]
d2 <- efc[, 2:3]
d3 <- efc[, 7:8]
d4 <- efc[, 10:12]

# show original
head(d1)

#>   c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4 c86cop5
#> 1      16        2      2     83      3       3       2       2       2       1
#> 2     148        2      2     88      3       3       3       3       3       4
#> 3      70        1      2     82      3       2       2       1       4       1
#> 4     168        1      2     67      4       4       1       3       1       1
#> 5     168        2      2     84      4       3       2       1       2       2
#> 6      16        2      2     85      4       2       2       3       3       3

library(sjlabelled)
# slightly change variables, to see effect
d2 <- as_label(d2)
d3 <- as_label(d3)

# replace duplicated columns, append remaining
replace_columns(d1, d2, d3, d4)

#> # A tibble: 908 x 12
#>    c12hour e15relat       e16sex e17age e42dep c82cop1 c83cop2  c84cop3 c85cop4
#>      <dbl> <fct>          <fct>   <dbl>  <dbl>   <dbl> <fct>    <fct>     <dbl>
#>  1      16 child          female     83      3       3 Sometim~ Someti~       2
#>  2     148 child          female     88      3       3 Often    Often         3
#>  3      70 spouse/partner female     82      3       2 Sometim~ Never         4
#>  4     168 spouse/partner female     67      4       4 Never    Often         1
#>  5     168 child          female     84      4       3 Sometim~ Never         2
#>  6      16 child          female     85      4       2 Sometim~ Often         3
#>  7     161 spouse/partner male       74      4       4 Sometim~ Always        1
#>  8     110 daughter or s~ female     87      4       3 Sometim~ Someti~       1
#>  9      28 child          female     79      4       3 Sometim~ Often         2
#> 10      40 child          female     83      4       3 Sometim~ Never         2
#> # ... with 898 more rows, and 3 more variables: c86cop5 <dbl>, c87cop6 <dbl>,
#> #   c88cop7 <dbl>

# replace duplicated columns, omit remaining
replace_columns(d1, d2, d3, d4, add.unique = FALSE)

#> # A tibble: 908 x 10
#>    c12hour e15relat       e16sex e17age e42dep c82cop1 c83cop2  c84cop3 c85cop4
#>      <dbl> <fct>          <fct>   <dbl>  <dbl>   <dbl> <fct>    <fct>     <dbl>
#>  1      16 child          female     83      3       3 Sometim~ Someti~       2
#>  2     148 child          female     88      3       3 Often    Often         3
#>  3      70 spouse/partner female     82      3       2 Sometim~ Never         4
#>  4     168 spouse/partner female     67      4       4 Never    Often         1
#>  5     168 child          female     84      4       3 Sometim~ Never         2
#>  6      16 child          female     85      4       2 Sometim~ Often         3
#>  7     161 spouse/partner male       74      4       4 Sometim~ Always        1
#>  8     110 daughter or s~ female     87      4       3 Sometim~ Someti~       1
#>  9      28 child          female     79      4       3 Sometim~ Often         2
#> 10      40 child          female     83      4       3 Sometim~ Never         2
#> # ... with 898 more rows, and 1 more variable: c86cop5 <dbl>