Dichotomize variables

Dichotomizes variables into dummy variables (0/1). Dichotomization is either done by median, mean or a specific value (see dich.by). dicho_if() is a scoped variant of dicho(), where recoding will be applied only to those variables that match the logical condition of predicate.

dicho(x, ..., dich.by = "median", as.num = FALSE, var.label = NULL,
  val.labels = NULL, append = TRUE, suffix = "_d")

dicho_if(x, predicate, dich.by = "median", as.num = FALSE,
  var.label = NULL, val.labels = NULL, append = TRUE, suffix = "_d")

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples' or package-vignette.

dich.by
Indicates the split criterion where a variable is dichotomized. Must be one of the following values (may be abbreviated):

"median" or "md"

by default, x is split into two groups at the median.

"mean" or "m"

splits x into two groups at the mean of x.

numeric value

splits x into two groups at the specific value. Note that the value is inclusive, i.e. dich.by = 10 will split x into one group with values from lowest to 10 and another group with values greater than 10.

as.num
Logical, if TRUE, return value will be numeric, not a factor.

var.label
Optional string, to set variable label attribute for the returned variable (see vignette Labelled Data and the sjlabelled-Package). If NULL (default), variable label attribute of x will be used (if present). If empty, variable label attributes will be removed.

val.labels
Optional character vector (of length two), to set value label attributes of dichotomized variable (see set_labels). If NULL (default), no value labels will be set.

append
Logical, if TRUE (the default) and x is a data frame, x including the new variables as additional columns is returned; if FALSE, only the new variables are returned.

suffix
String value, will be appended to variable (column) names of x, if x is a data frame. If x is not a data frame, this argument will be ignored. The default value to suffix column names in a data frame depends on the function call:

recoded variables (rec()) will be suffixed with "_r"

recoded variables (recode_to()) will be suffixed with "_r0"

dichotomized variables (dicho()) will be suffixed with "_d"

grouped variables (split_var()) will be suffixed with "_g"

grouped variables (group_var()) will be suffixed with "_gr"

standardized variables (std()) will be suffixed with "_z"

centered variables (center()) will be suffixed with "_c"

predicate
A predicate function to be applied to the columns. The variables for which predicate returns TRUE are selected.

Value

x, dichotomized. If x is a data frame, for append = TRUE, x including the dichotomized. variables as new columns is returned; if append = FALSE, only the dichotomized variables will be returned.

Details

dicho() also works on grouped data frames (see group_by). In this case, dichotomization is applied to the subsets of variables in x. See 'Examples'.

Note

Variable label attributes are preserved (unless changed via var.label-argument).

Examples

data(efc)
summary(efc$c12hour)

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
#>    4.00   10.00   20.00   42.40   42.75  168.00       6 

# split at median
table(dicho(efc$c12hour))

#> 
#>   0   1 
#> 464 438 

# split at mean
table(dicho(efc$c12hour, dich.by = "mean"))

#> 
#>   0   1 
#> 676 226 

# split between value lowest to 30, and above 30
table(dicho(efc$c12hour, dich.by = 30))

#> 
#>   0   1 
#> 606 296 

# sample data frame, values from 1-4
head(efc[, 6:10])

#>   c82cop1 c83cop2 c84cop3 c85cop4 c86cop5
#> 1       3       2       2       2       1
#> 2       3       3       3       3       4
#> 3       2       2       1       4       1
#> 4       4       1       3       1       1
#> 5       3       2       1       2       2
#> 6       2       2       3       3       3

# dichtomized values (1 to 2 = 0, 3 to 4 = 1)
library(dplyr)
efc %>%
  select(6:10) %>%
  dicho(dich.by = 2) %>%
  head()

#> # A tibble: 6 x 10
#>   c82cop1 c83cop2 c84cop3 c85cop4 c86cop5 c82cop1_d c83cop2_d c84cop3_d
#>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <fct>     <fct>     <fct>    
#> 1       3       2       2       2       1 1         0         0        
#> 2       3       3       3       3       4 1         1         1        
#> 3       2       2       1       4       1 0         0         0        
#> 4       4       1       3       1       1 1         0         1        
#> 5       3       2       1       2       2 1         0         0        
#> 6       2       2       3       3       3 0         0         1        
#> # ... with 2 more variables: c85cop4_d <fct>, c86cop5_d <fct>

# dichtomize several variables in a data frame
dicho(efc, c12hour, e17age, c160age, append = FALSE)

#> # A tibble: 908 x 3
#>    c12hour_d e17age_d c160age_d
#>    <fct>     <fct>    <fct>    
#>  1 0         1        1        
#>  2 1         1        0        
#>  3 1         1        1        
#>  4 1         0        1        
#>  5 1         1        0        
#>  6 0         1        1        
#>  7 1         0        1        
#>  8 1         1        1        
#>  9 1         0        1        
#> 10 1         1        0        
#> # ... with 898 more rows

# dichotomize and set labels
frq(dicho(
  efc, e42dep,
  var.label = "Dependency (dichotomized)",
  val.labels = c("lower", "higher"),
  append = FALSE
))

#> 
#> # Dependency (dichotomized) (e42dep_d) <categorical> 
#> # total N=908  valid N=901  mean=0.34  sd=0.47
#>  
#>  val  label frq raw.prc valid.prc cum.prc
#>    0  lower 597   65.75     66.26   66.26
#>    1 higher 304   33.48     33.74  100.00
#>   NA     NA   7    0.77        NA      NA
#> 
#> 

# works also with gouped data frames
mtcars %>%
  dicho(disp, append = FALSE) %>%
  table()

#> .
#>  0  1 
#> 16 16 

mtcars %>%
  group_by(cyl) %>%
  dicho(disp, append = FALSE) %>%
  table()

#> .
#>  1  2 
#> 18 14 

# dichotomizing grouped data frames leads to different
# results for a dichotomized variable, because the split
# value is different for each group.
# compare:
mtcars %>%
  group_by(cyl) %>%
  summarise(median = median(disp))

#> # A tibble: 3 x 2
#>     cyl median
#>   <dbl>  <dbl>
#> 1     4   108 
#> 2     6   168.
#> 3     8   350.

median(mtcars$disp)

#> [1] 196.3

# dichotomize only variables with more than 10 unique values
p <- function(x) dplyr::n_distinct(x) > 10
dicho_if(efc, predicate = p, append = FALSE)

#> # A tibble: 908 x 7
#>    c12hour_d e17age_d c160age_d barthtot_d neg_c_7_d pos_v_4_d quol_5_d
#>    <fct>     <fct>    <fct>     <fct>      <fct>     <fct>     <fct>   
#>  1 0         1        1         1          1         0         0       
#>  2 1         1        0         1          1         0         0       
#>  3 1         1        1         0          0         0         0       
#>  4 1         0        1         0          0         1         0       
#>  5 1         1        0         0          1         1         1       
#>  6 0         1        1         0          1         0         0       
#>  7 1         0        1         0          1         0         1       
#>  8 1         1        1         0          0         1         1       
#>  9 1         0        1         0          1         0         0       
#> 10 1         1        0         0          0         0         0       
#> # ... with 898 more rows