Split numeric variables into smaller groups

Recode numeric variables into equal sized groups, i.e. a variable is cut into a smaller number of groups at specific cut points. split_var_if() is a scoped variant of split_var(), where transformation will be applied only to those variables that match the logical condition of predicate.

split_var(x, ..., n, as.num = FALSE, val.labels = NULL,
  var.label = NULL, inclusive = FALSE, append = TRUE,
  suffix = "_g")

split_var_if(x, predicate, n, as.num = FALSE, val.labels = NULL,
  var.label = NULL, inclusive = FALSE, append = TRUE,
  suffix = "_g")

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples' or package-vignette.

n
The new number of groups that x should be split into.

as.num
Logical, if TRUE, return value will be numeric, not a factor.

val.labels
Optional character vector, to set value label attributes of recoded variable (see vignette Labelled Data and the sjlabelled-Package). If NULL (default), no value labels will be set. Value labels can also be directly defined in the rec-syntax, see 'Details'.

var.label
Optional string, to set variable label attribute for the returned variable (see vignette Labelled Data and the sjlabelled-Package). If NULL (default), variable label attribute of x will be used (if present). If empty, variable label attributes will be removed.

inclusive
Logical; if TRUE, cut point value are included in the preceeding group. This may be necessary if cutting a vector into groups does not define proper ("equal sized") group sizes. See 'Note' and 'Examples'.

append
Logical, if TRUE (the default) and x is a data frame, x including the new variables as additional columns is returned; if FALSE, only the new variables are returned.

suffix
String value, will be appended to variable (column) names of x, if x is a data frame. If x is not a data frame, this argument will be ignored. The default value to suffix column names in a data frame depends on the function call:

recoded variables (rec()) will be suffixed with "_r"

recoded variables (recode_to()) will be suffixed with "_r0"

dichotomized variables (dicho()) will be suffixed with "_d"

grouped variables (split_var()) will be suffixed with "_g"

grouped variables (group_var()) will be suffixed with "_gr"

standardized variables (std()) will be suffixed with "_z"

centered variables (center()) will be suffixed with "_c"

predicate
A predicate function to be applied to the columns. The variables for which predicate returns TRUE are selected.

Value

A grouped variable with equal sized groups. If x is a data frame, for append = TRUE, x including the grouped variables as new columns is returned; if append = FALSE, only the grouped variables will be returned.

Details

split_var() splits a variable into equal sized groups, where the amount of groups depends on the groupcount-argument. Thus, this functions cuts a variable into groups at the specified quantiles.

By contrast, group_var recodes a variable into groups, where groups have the same value range (e.g., from 1-5, 6-10, 11-15 etc.).

split_var() also works on grouped data frames (see group_by). In this case, splitting is applied to the subsets of variables in x. See 'Examples'.

Note

In case a vector has only few number of unique values, splitting into equal sized groups may fail. In this case, use the inclusive-argument to shift a value at the cut point into the lower, preceeding group to get equal sized groups. See 'Examples'.

See also

Examples

data(efc)
# non-grouped
table(efc$neg_c_7)

#> 
#>   7   8   9  10  11  12  13  14  15  16  17  18  19  20  21  22  23  24  25  27 
#>  75  99 106 120  96  85  64  54  45  30  35  26  16  16   2   7   4   3   6   1 
#>  28 
#>   2 

# split into 3 groups
table(split_var(efc$neg_c_7, n = 3))

#> 
#>   1   2   3 
#> 280 301 311 

# split multiple variables into 3 groups
split_var(efc, neg_c_7, pos_v_4, e17age, n = 3, append = FALSE)

#> # A tibble: 908 x 3
#>    neg_c_7_g pos_v_4_g e17age_g
#>    <fct>     <fct>     <fct>   
#>  1 2         2         3       
#>  2 3         1         3       
#>  3 2         2         2       
#>  4 2         3         1       
#>  5 2         3         3       
#>  6 3         1         3       
#>  7 3         2         1       
#>  8 2         3         3       
#>  9 3         2         2       
#> 10 2         2         3       
#> # ... with 898 more rows

frq(split_var(efc, neg_c_7, pos_v_4, e17age, n = 3, append = FALSE))

#> 
#> # Negative impact with 7 items (neg_c_7_g) <categorical> 
#> # total N=908  valid N=892  mean=2.03  sd=0.81
#>  
#>   val frq raw.prc valid.prc cum.prc
#>     1 280   30.84     31.39   31.39
#>     2 301   33.15     33.74   65.13
#>     3 311   34.25     34.87  100.00
#>  <NA>  16    1.76        NA      NA
#> 
#> 
#> # Positive value with 4 items (pos_v_4_g) <categorical> 
#> # total N=908  valid N=881  mean=2.03  sd=0.81
#>  
#>   val frq raw.prc valid.prc cum.prc
#>     1 275   30.29     31.21   31.21
#>     2 304   33.48     34.51   65.72
#>     3 302   33.26     34.28  100.00
#>  <NA>  27    2.97        NA      NA
#> 
#> 
#> # elder' age (e17age_g) <categorical> 
#> # total N=908  valid N=891  mean=2.05  sd=0.82
#>  
#>   val frq raw.prc valid.prc cum.prc
#>     1 274   30.18     30.75   30.75
#>     2 294   32.38     33.00   63.75
#>     3 323   35.57     36.25  100.00
#>  <NA>  17    1.87        NA      NA
#> 
#> 

# original
table(efc$e42dep)

#> 
#>   1   2   3   4 
#>  66 225 306 304 

# two groups, non-inclusive cut-point
# vector split leads to unequal group sizes
table(split_var(efc$e42dep, n = 2))

#> 
#>   1   2 
#> 291 610 

# two groups, inclusive cut-point
# group sizes are equal
table(split_var(efc$e42dep, n = 2, inclusive = TRUE))

#> 
#>   1   2 
#> 597 304 

# Unlike dplyr's ntile(), split_var() never splits a value
# into two different categories, i.e. you always get a clean
# separation of original categories
library(dplyr)

x <- dplyr::ntile(efc$neg_c_7, n = 3)
table(efc$neg_c_7, x)

#>     x
#>        1   2   3
#>   7   75   0   0
#>   8   99   0   0
#>   9  106   0   0
#>   10  18 102   0
#>   11   0  96   0
#>   12   0  85   0
#>   13   0  14  50
#>   14   0   0  54
#>   15   0   0  45
#>   16   0   0  30
#>   17   0   0  35
#>   18   0   0  26
#>   19   0   0  16
#>   20   0   0  16
#>   21   0   0   2
#>   22   0   0   7
#>   23   0   0   4
#>   24   0   0   3
#>   25   0   0   6
#>   27   0   0   1
#>   28   0   0   2

x <- split_var(efc$neg_c_7, n = 3)
table(efc$neg_c_7, x)

#>     x
#>        1   2   3
#>   7   75   0   0
#>   8   99   0   0
#>   9  106   0   0
#>   10   0 120   0
#>   11   0  96   0
#>   12   0  85   0
#>   13   0   0  64
#>   14   0   0  54
#>   15   0   0  45
#>   16   0   0  30
#>   17   0   0  35
#>   18   0   0  26
#>   19   0   0  16
#>   20   0   0  16
#>   21   0   0   2
#>   22   0   0   7
#>   23   0   0   4
#>   24   0   0   3
#>   25   0   0   6
#>   27   0   0   1
#>   28   0   0   2

# works also with gouped data frames
mtcars %>%
  split_var(disp, n = 3, append = FALSE) %>%
  table()

#> .
#>  1  2  3 
#> 11 10 11 

mtcars %>%
  group_by(cyl) %>%
  split_var(disp, n = 3, append = FALSE) %>%
  table()

#> .
#>  1  2  3 
#> 10  8 14