Standardize and center variables

std() computes a z-transformation (standardized and centered) on the input. center() centers the input. std_if() and center_if() are scoped variants of std() and center(), where transformation will be applied only to those variables that match the logical condition of predicate.

std(x, ..., robust = c("sd", "2sd", "gmd", "mad"), include.fac = FALSE,
  append = TRUE, suffix = "_z")

std_if(x, predicate, robust = c("sd", "2sd", "gmd", "mad"),
  include.fac = FALSE, append = TRUE, suffix = "_z")

center(x, ..., include.fac = FALSE, append = TRUE, suffix = "_c")

center_if(x, predicate, include.fac = FALSE, append = TRUE, suffix = "_c")

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples' or package-vignette.

robust
Character vector, indicating the method applied when standardizing variables with std(). By default, standardization is achieved by dividing the centered variables by their standard deviation (robust = "sd"). However, for skewed distributions, the median absolute deviation (MAD, robust = "mad") or Gini's mean difference (robust = "gmd") might be more robust measures of dispersion. For the latter option, sjstats needs to be installed. robust = "2sd" divides the centered variables by two standard deviations, following a suggestion by Gelman (2008), so the rescaled input is comparable to binary variables.

include.fac
Logical, if TRUE, factors will be converted to numeric vectors and also standardized or centered.

append
Logical, if TRUE (the default) and x is a data frame, x including the new variables as additional columns is returned; if FALSE, only the new variables are returned.

suffix
String value, will be appended to variable (column) names of x, if x is a data frame. If x is not a data frame, this argument will be ignored. The default value to suffix column names in a data frame depends on the function call:

recoded variables (rec()) will be suffixed with "_r"

recoded variables (recode_to()) will be suffixed with "_r0"

dichotomized variables (dicho()) will be suffixed with "_d"

grouped variables (split_var()) will be suffixed with "_g"

grouped variables (group_var()) will be suffixed with "_gr"

standardized variables (std()) will be suffixed with "_z"

centered variables (center()) will be suffixed with "_c"

predicate
A predicate function to be applied to the columns. The variables for which predicate returns TRUE are selected.

Value

If x is a vector, returns a vector with standardized or centered variables. If x is a data frame, for append = TRUE, x including the transformed variables as new columns is returned; if append = FALSE, only the transformed variables will be returned.

Details

std() and center() also work on grouped data frames (see group_by). In this case, standardization or centering is applied to the subsets of variables in x. See 'Examples'.

For more complicated models with many predictors, Gelman and Hill (2007) suggest leaving binary inputs as is and only standardize continuous predictors by dividing by two standard deviations. This ensures a rough comparability in the coefficients.

Note

std() and center() only return a vector, if x is a vector. If x is a data frame and only one variable is specified in the ...-ellipses argument, both functions do return a data frame (see 'Examples').

References

Gelman A (2008) Scaling regression inputs by dividing by two standard deviations. Statistics in Medicine 27: 2865–2873. http://www.stat.columbia.edu/~gelman/research/published/standardizing7.pdf

Gelman A, Hill J (2007) Data Analysis Using Regression and Multilevel/Hierarchical Models. Cambdridge, Cambdrige University Press: 55-57

Examples

data(efc)
std(efc$c160age) %>% head()

#> [1]  0.19002509  0.04023278  1.98753275  1.16367507 -0.48404028  0.19002509

std(efc, e17age, c160age, append = FALSE) %>% head()

#> # A tibble: 6 x 2
#>   e17age_z c160age_z
#>      <dbl>     <dbl>
#> 1    0.479    0.190 
#> 2    1.10     0.0402
#> 3    0.356    1.99  
#> 4   -1.50     1.16  
#> 5    0.603   -0.484 
#> 6    0.726    0.190 

center(efc$c160age) %>% head()

#> [1]  2.5371809  0.5371809 26.5371809 15.5371809 -6.4628191  2.5371809

center(efc, e17age, c160age, append = FALSE) %>% head()

#> # A tibble: 6 x 2
#>   e17age_c c160age_c
#>      <dbl>     <dbl>
#> 1     3.88     2.54 
#> 2     8.88     0.537
#> 3     2.88    26.5  
#> 4   -12.1     15.5  
#> 5     4.88    -6.46 
#> 6     5.88     2.54 

# NOTE!
std(efc$e17age) # returns a vector

#>   [1]  0.47919925  1.09691703  0.35565569 -1.49749765  0.60274280  0.72628636
#>   [7] -0.63269276  0.97337347 -0.01497498  0.47919925 -1.37395409  2.20880903
#>  [13]  0.10856858 -0.50914920  0.35565569  1.22046058  0.10856858 -0.87977987
#>  [19]  1.83817836 -0.01497498  1.22046058 -1.49749765  1.83817836  0.47919925
#>  [25]  0.72628636  0.10856858  1.09691703 -0.38560564  0.60274280  1.96172192
#>  [31]  1.09691703  0.97337347  1.22046058  0.10856858 -0.50914920  0.35565569
#>  [37] -1.25041054  1.46754769  0.84982991  0.84982991  0.60274280 -1.25041054
#>  [43] -1.49749765 -1.49749765 -1.62104120 -0.01497498 -0.87977987 -1.74458476
#>  [49] -0.50914920  0.97337347 -1.37395409 -0.50914920 -1.74458476 -1.49749765
#>  [55]  0.23211214  0.47919925          NA  0.35565569 -0.01497498 -0.87977987
#>  [61]  0.97337347  0.72628636  1.09691703 -0.63269276  0.97337347  1.09691703
#>  [67] -0.38560564  0.23211214  0.10856858 -1.37395409  0.35565569          NA
#>  [73]  1.46754769 -1.62104120  1.22046058 -0.01497498 -1.25041054  1.59109125
#>  [79] -0.50914920 -0.38560564 -0.26206209          NA -0.38560564 -0.38560564
#>  [85] -0.01497498 -0.26206209 -1.49749765 -1.62104120 -0.13851853  0.60274280
#>  [91] -0.01497498 -0.26206209 -1.62104120  1.83817836 -1.74458476  1.34400414
#>  [97] -1.62104120  1.34400414  0.72628636 -0.26206209 -0.38560564 -1.62104120
#> [103] -0.38560564  0.10856858  1.22046058 -1.25041054 -0.50914920 -0.50914920
#> [109]  0.10856858  0.10856858 -0.63269276  0.35565569  1.46754769  0.10856858
#> [115] -0.63269276  1.34400414  0.47919925  0.35565569 -0.75623631  0.60274280
#> [121]  1.34400414 -1.74458476  0.60274280 -0.87977987  1.96172192 -1.49749765
#> [127]  0.23211214  1.59109125  0.47919925 -0.63269276  1.46754769 -0.13851853
#> [133]  0.23211214 -0.38560564  1.71463481  2.20880903  1.71463481  0.23211214
#> [139] -0.50914920  0.60274280 -0.13851853  0.60274280  0.97337347  0.47919925
#> [145]  1.83817836  0.72628636 -1.12686698 -1.25041054 -0.38560564  0.47919925
#> [151]  1.46754769  1.34400414  0.97337347  1.46754769 -0.87977987 -0.75623631
#> [157] -0.01497498  0.60274280  0.35565569 -1.74458476 -0.13851853  1.83817836
#> [163]  1.46754769          NA  0.10856858 -1.12686698  0.60274280  0.35565569
#> [169] -0.38560564 -0.26206209 -1.62104120  1.34400414 -0.38560564 -1.25041054
#> [175]  1.09691703 -1.00332342  1.71463481  2.20880903  0.60274280 -0.01497498
#> [181]  1.46754769  0.60274280 -0.50914920  0.23211214 -0.26206209 -0.50914920
#> [187] -0.63269276  1.59109125 -0.01497498 -0.63269276 -0.38560564          NA
#> [193]  0.60274280 -1.74458476  0.72628636 -0.50914920  1.34400414  0.47919925
#> [199]  0.97337347  0.84982991  1.71463481  1.59109125  1.59109125  0.60274280
#> [205]  0.47919925  0.35565569  0.97337347  1.22046058  1.46754769  0.60274280
#> [211]  1.71463481  0.72628636 -0.01497498  0.60274280 -1.74458476  1.22046058
#> [217]  0.60274280  0.23211214 -0.26206209 -1.74458476  1.71463481  2.08526547
#> [223]  1.34400414  0.84982991  1.09691703 -0.01497498  1.46754769  0.60274280
#> [229] -1.00332342 -1.49749765  0.97337347  0.60274280  0.97337347  0.60274280
#> [235]  0.84982991  0.47919925 -1.25041054 -0.50914920 -0.01497498 -0.38560564
#> [241] -0.38560564  0.35565569  1.09691703  1.34400414  0.97337347 -0.38560564
#> [247] -0.38560564  0.84982991  0.47919925  1.22046058  0.35565569 -0.01497498
#> [253] -1.49749765 -0.75623631  1.22046058  1.96172192 -1.74458476  1.22046058
#> [259]  0.47919925  0.84982991  0.35565569  1.71463481  1.34400414 -0.01497498
#> [265]  2.20880903 -0.13851853 -0.63269276  0.84982991  1.09691703 -0.26206209
#> [271] -0.38560564 -0.63269276 -0.38560564 -1.00332342 -0.63269276  0.35565569
#> [277]  0.35565569 -0.63269276 -0.75623631 -1.49749765 -1.12686698 -1.12686698
#> [283] -1.12686698 -1.49749765  0.10856858 -1.25041054  1.22046058  0.47919925
#> [289]  0.47919925 -1.49749765  0.60274280  1.34400414  0.10856858  0.23211214
#> [295]  0.97337347 -0.75623631 -1.00332342  0.10856858 -1.25041054 -0.75623631
#> [301]  1.09691703 -0.75623631 -0.13851853  0.84982991 -0.50914920  1.22046058
#> [307]  0.47919925 -0.13851853 -1.25041054 -1.12686698  0.84982991 -1.49749765
#> [313] -0.63269276  0.97337347 -1.62104120 -1.37395409 -0.87977987  0.10856858
#> [319]  1.22046058  0.60274280  1.71463481  0.84982991  0.60274280  0.84982991
#> [325]  0.47919925  0.84982991  1.09691703 -0.50914920 -0.38560564          NA
#> [331]  0.23211214 -0.13851853  0.47919925  1.83817836  0.60274280 -1.00332342
#> [337]  0.84982991  1.22046058  1.71463481 -0.50914920 -0.75623631 -1.12686698
#> [343] -0.50914920          NA -0.50914920  2.95007036 -1.12686698  0.60274280
#> [349] -0.13851853  0.72628636  0.23211214  1.34400414 -0.38560564 -0.50914920
#> [355] -0.63269276 -0.50914920  1.22046058 -0.75623631 -0.01497498  0.47919925
#> [361]  1.46754769  0.23211214 -0.26206209 -1.74458476 -0.75623631 -0.75623631
#> [367] -1.25041054 -0.01497498  0.35565569 -0.38560564  0.60274280  0.84982991
#> [373]  1.09691703 -0.38560564  1.83817836 -0.13851853 -1.49749765  0.23211214
#> [379] -1.25041054 -0.26206209 -0.75623631 -1.12686698  0.35565569  1.71463481
#> [385]  0.23211214 -1.12686698 -0.13851853  0.60274280  0.47919925  0.35565569
#> [391]  0.72628636  0.47919925 -1.37395409 -1.12686698 -0.38560564 -0.75623631
#> [397] -0.38560564  0.10856858 -1.49749765 -0.26206209 -0.01497498 -0.87977987
#> [403]  1.34400414  1.34400414 -1.62104120  0.10856858  0.23211214  0.72628636
#> [409]  0.47919925  0.72628636  0.60274280 -1.74458476 -1.00332342 -1.37395409
#> [415] -1.74458476 -0.75623631 -1.37395409 -1.49749765 -0.75623631  0.60274280
#> [421]  2.33235258 -0.13851853  0.35565569  1.22046058  0.60274280 -1.12686698
#> [427] -0.13851853  0.35565569  1.83817836  1.22046058 -0.75623631  0.10856858
#> [433]  0.60274280 -0.13851853  0.47919925 -0.26206209  0.35565569  0.23211214
#> [439] -0.87977987  1.09691703  0.10856858 -0.01497498 -1.62104120  0.84982991
#> [445]  0.72628636 -0.01497498  0.60274280          NA  0.97337347 -1.00332342
#> [451] -1.00332342 -0.38560564  0.10856858  0.23211214  1.34400414  0.35565569
#> [457] -0.26206209  1.96172192  0.84982991  0.35565569  1.34400414 -0.01497498
#> [463]  1.96172192  0.97337347 -0.13851853  0.47919925 -0.50914920  0.10856858
#> [469] -0.87977987 -0.38560564  0.60274280 -1.62104120 -0.26206209  0.60274280
#> [475] -1.37395409 -0.13851853 -1.74458476 -0.87977987 -1.49749765 -0.13851853
#> [481] -1.49749765  0.35565569 -0.50914920  0.35565569 -0.01497498 -0.38560564
#> [487]  1.22046058  1.59109125 -1.62104120  1.22046058 -0.01497498 -0.75623631
#> [493]  1.22046058 -0.01497498          NA  0.23211214  0.60274280 -0.26206209
#> [499] -0.75623631 -0.01497498 -0.75623631  0.10856858 -0.75623631  1.09691703
#> [505]  0.35565569  1.71463481 -1.12686698 -1.25041054 -1.12686698  0.84982991
#> [511] -0.50914920  0.23211214 -1.74458476  0.84982991  1.46754769 -0.87977987
#> [517]  0.97337347 -0.26206209 -1.74458476  0.23211214 -0.87977987  1.71463481
#> [523] -1.74458476 -0.26206209 -0.26206209 -0.38560564  0.72628636 -0.01497498
#> [529]  0.35565569  0.47919925  0.84982991 -1.00332342 -0.50914920  0.47919925
#> [535] -0.13851853 -0.63269276  1.09691703  0.84982991  0.35565569  1.22046058
#> [541]  1.46754769  1.34400414 -0.26206209 -1.74458476 -1.25041054 -1.37395409
#> [547] -1.12686698  0.35565569 -0.26206209  0.35565569 -1.25041054  0.23211214
#> [553]  0.84982991  0.72628636 -1.49749765 -1.12686698 -0.13851853  0.47919925
#> [559]  0.23211214 -0.63269276 -0.13851853  0.47919925 -1.25041054  2.20880903
#> [565] -1.12686698  1.59109125  1.46754769 -1.74458476          NA  1.83817836
#> [571] -1.25041054 -0.13851853  0.97337347  0.35565569  0.60274280 -0.01497498
#> [577]  0.72628636 -1.74458476 -0.38560564 -1.74458476 -0.13851853  1.46754769
#> [583]  1.09691703  0.47919925 -0.75623631 -0.01497498  0.84982991  0.35565569
#> [589] -0.50914920 -1.62104120  1.71463481  0.84982991  0.84982991 -1.25041054
#> [595] -1.62104120  0.47919925 -1.74458476 -1.00332342  0.47919925  1.59109125
#> [601]  0.47919925  1.96172192 -0.38560564  0.23211214 -0.01497498  0.97337347
#> [607] -0.63269276 -1.12686698  0.10856858 -0.50914920 -1.37395409 -1.25041054
#> [613] -1.37395409 -1.12686698  1.22046058          NA  0.47919925 -1.12686698
#> [619]  0.35565569  0.60274280  0.35565569  0.10856858  0.84982991 -1.25041054
#> [625]  0.97337347 -1.00332342 -1.62104120  0.72628636  0.72628636  1.96172192
#> [631] -0.38560564 -1.25041054 -1.37395409  0.35565569  1.59109125 -1.62104120
#> [637] -0.63269276 -0.50914920  1.34400414 -0.87977987  1.59109125 -1.12686698
#> [643] -0.01497498 -0.01497498 -1.25041054 -1.12686698  0.35565569 -0.01497498
#> [649] -0.87977987 -0.38560564 -1.74458476  1.59109125 -1.12686698  1.22046058
#> [655] -0.87977987 -0.50914920 -0.75623631  0.35565569  0.10856858  0.35565569
#> [661]  0.35565569 -1.00332342 -0.26206209 -0.75623631  0.10856858 -1.00332342
#> [667] -1.74458476  0.47919925 -1.49749765 -1.37395409 -1.00332342 -0.26206209
#> [673] -0.87977987 -1.37395409 -0.26206209 -1.12686698  0.60274280  1.22046058
#> [679] -1.37395409  1.22046058 -1.37395409 -0.87977987 -0.13851853  0.60274280
#> [685]  1.59109125 -1.37395409 -1.62104120 -1.00332342 -1.74458476  0.23211214
#> [691]  0.35565569 -0.50914920  0.10856858 -1.12686698  1.46754769  1.83817836
#> [697] -1.62104120  0.97337347  0.84982991  0.47919925 -0.75623631 -1.12686698
#> [703] -0.75623631  0.47919925 -0.38560564  1.34400414 -0.87977987 -0.13851853
#> [709] -1.00332342 -0.87977987  0.84982991 -1.12686698 -1.49749765 -1.74458476
#> [715] -0.63269276 -0.38560564 -0.01497498 -1.62104120 -0.75623631  1.34400414
#> [721] -1.37395409 -1.49749765  0.84982991 -0.01497498 -0.50914920  0.72628636
#> [727] -0.26206209 -0.01497498  1.71463481  0.10856858 -0.26206209  0.23211214
#> [733]  0.10856858  0.84982991 -1.49749765  1.46754769 -0.87977987  0.35565569
#> [739]  1.83817836  1.22046058  0.47919925 -0.50914920  1.22046058  0.10856858
#> [745]  0.60274280 -0.26206209 -0.50914920 -0.01497498  0.72628636  0.97337347
#> [751]  0.35565569  1.34400414  0.47919925  0.72628636  0.10856858 -1.62104120
#> [757] -0.01497498  1.46754769  1.22046058 -1.25041054 -0.75623631  0.47919925
#> [763] -0.01497498  0.72628636  0.23211214  0.72628636 -1.62104120 -0.63269276
#> [769] -0.75623631 -0.38560564  0.47919925 -0.13851853  1.09691703  1.22046058
#> [775]  0.60274280 -0.75623631  1.59109125 -1.74458476 -1.25041054 -0.26206209
#> [781] -0.01497498 -0.75623631 -1.37395409 -1.25041054  0.10856858  0.23211214
#> [787] -1.49749765 -0.63269276  1.09691703 -1.12686698 -0.75623631 -1.49749765
#> [793] -0.63269276  0.97337347  1.46754769 -0.38560564  0.97337347 -0.87977987
#> [799] -0.26206209  1.22046058  0.60274280 -0.13851853  0.23211214 -0.63269276
#> [805]  0.10856858 -0.26206209 -0.01497498  1.22046058  1.34400414  0.97337347
#> [811]  0.23211214  0.35565569 -1.00332342 -0.50914920 -1.25041054 -0.26206209
#> [817]  0.23211214  0.97337347 -0.75623631 -1.74458476 -1.37395409 -0.87977987
#> [823] -1.12686698 -1.37395409 -1.12686698 -1.37395409 -1.25041054 -0.13851853
#> [829] -0.63269276 -1.37395409  0.84982991 -0.01497498 -0.50914920  0.72628636
#> [835] -1.00332342 -0.63269276 -0.63269276  0.84982991 -1.62104120 -1.12686698
#> [841]  0.84982991  0.72628636  0.35565569 -0.01497498 -1.25041054 -1.49749765
#> [847]  1.83817836 -0.50914920 -0.01497498 -0.13851853 -1.62104120 -1.74458476
#> [853]  1.09691703 -0.13851853  0.47919925  0.97337347  2.45589614 -1.62104120
#> [859]  1.22046058 -1.00332342  0.35565569 -1.25041054  0.23211214  0.60274280
#> [865]  1.34400414  0.10856858 -0.01497498 -1.49749765 -1.00332342 -0.50914920
#> [871] -1.37395409 -0.50914920 -0.75623631  0.23211214 -0.38560564 -1.74458476
#> [877]  0.47919925 -0.01497498 -0.75623631 -0.13851853  0.47919925 -0.63269276
#> [883]  0.23211214  0.97337347 -1.49749765 -0.01497498  0.23211214 -1.12686698
#> [889] -1.74458476  1.71463481 -0.63269276 -0.26206209 -0.50914920  1.34400414
#> [895]  1.46754769  0.84982991 -1.74458476  0.35565569  0.10856858 -0.63269276
#> [901] -1.74458476 -1.49749765          NA          NA          NA          NA
#> [907]          NA          NA
#> attr(,"label")
#> [1] "elder' age"

std(efc, e17age) # returns a tibble

#> # A tibble: 908 x 27
#>    c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4
#>      <dbl>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1      16        2      2     83      3       3       2       2       2
#>  2     148        2      2     88      3       3       3       3       3
#>  3      70        1      2     82      3       2       2       1       4
#>  4     168        1      2     67      4       4       1       3       1
#>  5     168        2      2     84      4       3       2       1       2
#>  6      16        2      2     85      4       2       2       3       3
#>  7     161        1      1     74      4       4       2       4       1
#>  8     110        4      2     87      4       3       2       2       1
#>  9      28        2      2     79      4       3       2       3       2
#> 10      40        2      2     83      4       3       2       1       2
#> # ... with 898 more rows, and 18 more variables: c86cop5 <dbl>, c87cop6 <dbl>,
#> #   c88cop7 <dbl>, c89cop8 <dbl>, c90cop9 <dbl>, c160age <dbl>, c161sex <dbl>,
#> #   c172code <dbl>, c175empl <dbl>, barthtot <dbl>, neg_c_7 <dbl>,
#> #   pos_v_4 <dbl>, quol_5 <dbl>, resttotn <dbl>, tot_sc_e <dbl>, n4pstu <dbl>,
#> #   nur_pst <dbl>, e17age_z <dbl>

# works with mutate()
library(dplyr)
efc %>%
  select(e17age, neg_c_7) %>%
  mutate(age_std = std(e17age), burden = center(neg_c_7)) %>%
  head()

#>   e17age neg_c_7    age_std     burden
#> 1     83      12  0.4791992  0.1502242
#> 2     88      20  1.0969170  8.1502242
#> 3     82      11  0.3556557 -0.8497758
#> 4     67      10 -1.4974976 -1.8497758
#> 5     84      12  0.6027428  0.1502242
#> 6     85      19  0.7262864  7.1502242

# works also with grouped data frames
mtcars %>% std(disp)

#> # A tibble: 32 x 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb  disp_z
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4 -0.571 
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4 -0.571 
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1 -0.990 
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1  0.220 
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2  1.04  
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1 -0.0462
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4  1.04  
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2 -0.678 
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2 -0.726 
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4 -0.509 
#> # ... with 22 more rows

# compare new column "disp_z" w/ output above
mtcars %>%
  group_by(cyl) %>%
  std(disp)

#> # A tibble: 32 x 12
#> # Groups:   cyl [3]
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb disp_z
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4 -0.561
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4 -0.561
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1  0.107
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1  1.80 
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2  0.102
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1  1.00 
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4  0.102
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2  1.55 
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2  1.33 
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4 -0.378
#> # ... with 22 more rows

data(iris)
# also standardize factors
std(iris, include.fac = TRUE, append = FALSE)

#> # A tibble: 150 x 5
#>    Sepal.Length_z Sepal.Width_z Petal.Length_z Petal.Width_z Species_z
#>             <dbl>         <dbl>          <dbl>         <dbl>     <dbl>
#>  1         -0.898        1.02            -1.34         -1.31     -1.22
#>  2         -1.14        -0.132           -1.34         -1.31     -1.22
#>  3         -1.38         0.327           -1.39         -1.31     -1.22
#>  4         -1.50         0.0979          -1.28         -1.31     -1.22
#>  5         -1.02         1.25            -1.34         -1.31     -1.22
#>  6         -0.535        1.93            -1.17         -1.05     -1.22
#>  7         -1.50         0.786           -1.34         -1.18     -1.22
#>  8         -1.02         0.786           -1.28         -1.31     -1.22
#>  9         -1.74        -0.361           -1.34         -1.31     -1.22
#> 10         -1.14         0.0979          -1.28         -1.44     -1.22
#> # ... with 140 more rows

# don't standardize factors
std(iris, include.fac = FALSE, append = FALSE)

#> # A tibble: 150 x 5
#>    Sepal.Length_z Sepal.Width_z Petal.Length_z Petal.Width_z Species_z
#>             <dbl>         <dbl>          <dbl>         <dbl> <fct>    
#>  1         -0.898        1.02            -1.34         -1.31 setosa   
#>  2         -1.14        -0.132           -1.34         -1.31 setosa   
#>  3         -1.38         0.327           -1.39         -1.31 setosa   
#>  4         -1.50         0.0979          -1.28         -1.31 setosa   
#>  5         -1.02         1.25            -1.34         -1.31 setosa   
#>  6         -0.535        1.93            -1.17         -1.05 setosa   
#>  7         -1.50         0.786           -1.34         -1.18 setosa   
#>  8         -1.02         0.786           -1.28         -1.31 setosa   
#>  9         -1.74        -0.361           -1.34         -1.31 setosa   
#> 10         -1.14         0.0979          -1.28         -1.44 setosa   
#> # ... with 140 more rows

# standardize only variables with more than 10 unique values
p <- function(x) dplyr::n_distinct(x) > 10
std_if(efc, predicate = p, append = FALSE)

#> # A tibble: 908 x 7
#>    c12hour_z e17age_z c160age_z barthtot_z neg_c_7_z pos_v_4_z quol_5_z
#>        <dbl>    <dbl>     <dbl>      <dbl>     <dbl>     <dbl>    <dbl>
#>  1   -0.520    0.479     0.190       0.354    0.0389    -0.213  -0.0695
#>  2    2.08     1.10      0.0402      0.354    2.11      -0.658  -0.823 
#>  3    0.543    0.356     1.99       -1.00    -0.220      0.233  -1.39  
#>  4    2.47    -1.50      1.16       -2.19    -0.479      1.12   -0.446 
#>  5    2.47     0.603    -0.484      -1.34     0.0389     1.12    0.872 
#>  6   -0.520    0.726     0.190      -0.154    1.85      -1.55   -1.20  
#>  7    2.33    -0.633     0.565      -2.02     0.816      0.233   1.06  
#>  8    1.33     0.973     1.01       -1.00    -0.220      0.679   1.06  
#>  9   -0.283   -0.0150    0.415      -1.68     0.816      0.233  -1.20  
#> 10   -0.0472   0.479    -0.334      -2.19    -0.479      0.233   0.119 
#> # ... with 898 more rows