Convert variable into character vector and replace values with associated value labels

This function converts (replaces) variable values (also of factors or character vectors) with their associated value labels and returns them as character vector. This is just a convenient wrapper for as.character(to_label(x)).

to_character(x, ..., add.non.labelled = FALSE, prefix = FALSE,
  var.label = NULL, drop.na = TRUE, drop.levels = FALSE)

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples' or package-vignette.

add.non.labelled
Logical, if TRUE, values without associated value label will also be converted to labels (as is). See 'Examples'.

prefix
Logical, if TRUE, the value labels used as factor levels or character values will be prefixed with their associated values. See 'Examples'.

var.label
Optional string, to set variable label attribute for the returned variable (see vignette Labelled Data and the sjlabelled-Package). If NULL (default), variable label attribute of x will be used (if present). If empty, variable label attributes will be removed.

drop.na
Logical, if TRUE, tagged NA values with value labels will be converted to regular NA's. Else, tagged NA values will be replaced with their value labels. See 'Examples' and get_na.

drop.levels
Logical, if TRUE, unused factor levels will be dropped (i.e. droplevels will be applied before returning the result).

Value

A character vector with the associated value labels as values. If x is a data frame, the complete data frame x will be returned, where variables specified in ... are coerced to character variables; if ... is not specified, applies to all variables in the data frame.

Note

Value labels will be removed when converting variables to factors, variable labels, however, are preserved.

This function is kept for backwards-compatibility. It is preferred to use as_character.

Examples

library(sjlabelled)
data(efc)
print(get_labels(efc)['c161sex'])

#> $c161sex
#> [1] "Male"   "Female"
#> 

head(efc$c161sex)

#> [1] 2 2 1 1 2 1

head(to_character(efc$c161sex))

#> [1] "Female" "Female" "Male"   "Male"   "Female" "Male"  

print(get_labels(efc)['e42dep'])

#> $e42dep
#> [1] "independent"          "slightly dependent"   "moderately dependent"
#> [4] "severely dependent"  
#> 

table(efc$e42dep)

#> 
#>   1   2   3   4 
#>  66 225 306 304 

table(to_character(efc$e42dep))

#> 
#>          independent moderately dependent   severely dependent 
#>                   66                  306                  304 
#>   slightly dependent 
#>                  225 

head(efc$e42dep)

#> [1] 3 3 3 4 4 4

head(to_character(efc$e42dep))

#> [1] "moderately dependent" "moderately dependent" "moderately dependent"
#> [4] "severely dependent"   "severely dependent"   "severely dependent"  

# numeric values w/o value labels will also be converted into character
str(efc$e17age)

#>  num [1:908] 83 88 82 67 84 85 74 87 79 83 ...
#>  - attr(*, "label")= chr "elder' age"

str(to_character(efc$e17age))

#>  chr [1:908] "83" "88" "82" "67" "84" "85" "74" "87" "79" "83" "68" "97" ...
#>  - attr(*, "label")= chr "elder' age"

# factor with non-numeric levels, non-prefixed and prefixed
x <- factor(c("a", "b", "c"))
x <- set_labels(x, labels = c("ape", "bear", "cat"))

to_character(x, prefix = FALSE)

#> [1] "ape"  "bear" "cat" 

to_character(x, prefix = TRUE)

#> [1] "[a] ape"  "[b] bear" "[c] cat" 

# create vector
x <- c(1, 2, 3, 2, 4, NA)
# add less labels than values
x <- set_labels(x,
                labels = c("yes", "maybe", "no"),
                force.labels = FALSE,
                force.values = FALSE)

#> "x" has more values than "labels", hence not all values are labelled.

# convert to character w/o non-labelled values
to_character(x)

#> [1] "yes"   "maybe" "no"    "maybe" NA      NA     

# convert to character, including non-labelled values
to_character(x, add.non.labelled = TRUE)

#> [1] "yes"   "maybe" "no"    "maybe" "4"     NA     

# create labelled integer, with missing flag
library(haven)
x <- labelled(c(1:3, tagged_na("a", "c", "z"), 4:1, 2:3),
              c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
                "Refused" = tagged_na("a"), "Not home" = tagged_na("z")))
# to character, with missing labels
to_character(x, drop.na = FALSE)

#>  [1] "Agreement"    NA             NA             "Refused"      "First"       
#>  [6] "Not home"     "Disagreement" NA             NA             "Agreement"   
#> [11] NA             NA            

# to character, missings removed
to_character(x, drop.na = TRUE)

#>  [1] "Agreement"    NA             NA             NA             NA            
#>  [6] NA             "Disagreement" NA             NA             "Agreement"   
#> [11] NA             NA            

# keep missings, and use non-labelled values as well
to_character(x, add.non.labelled = TRUE, drop.na = FALSE)

#>  [1] "Agreement"    "2"            "3"            "Refused"      "First"       
#>  [6] "Not home"     "Disagreement" "3"            "2"            "Agreement"   
#> [11] "2"            "3"           

# easily coerce specific variables in a data frame to character
# and keep other variables, with their class preserved
to_character(efc, e42dep, e16sex, c172code)

#> # A tibble: 908 x 26
#>    c12hour e15relat e16sex e17age e42dep        c82cop1 c83cop2 c84cop3 c85cop4
#>      <dbl>    <dbl> <chr>   <dbl> <chr>           <dbl>   <dbl>   <dbl>   <dbl>
#>  1      16        2 female     83 moderately d~       3       2       2       2
#>  2     148        2 female     88 moderately d~       3       3       3       3
#>  3      70        1 female     82 moderately d~       2       2       1       4
#>  4     168        1 female     67 severely dep~       4       1       3       1
#>  5     168        2 female     84 severely dep~       3       2       1       2
#>  6      16        2 female     85 severely dep~       2       2       3       3
#>  7     161        1 male       74 severely dep~       4       2       4       1
#>  8     110        4 female     87 severely dep~       3       2       2       1
#>  9      28        2 female     79 severely dep~       3       2       3       2
#> 10      40        2 female     83 severely dep~       3       2       1       2
#> # ... with 898 more rows, and 17 more variables: c86cop5 <dbl>, c87cop6 <dbl>,
#> #   c88cop7 <dbl>, c89cop8 <dbl>, c90cop9 <dbl>, c160age <dbl>, c161sex <dbl>,
#> #   c172code <chr>, c175empl <dbl>, barthtot <dbl>, neg_c_7 <dbl>,
#> #   pos_v_4 <dbl>, quol_5 <dbl>, resttotn <dbl>, tot_sc_e <dbl>, n4pstu <dbl>,
#> #   nur_pst <dbl>