Basic descriptive statistics

This function prints a basic descriptive statistic, including variable labels.

descr(
  x,
  ...,
  max.length = NULL,
  weights = NULL,
  show = "all",
  out = c("txt", "viewer", "browser"),
  encoding = "UTF-8",
  file = NULL
)

Arguments

x
A vector or a data frame. May also be a grouped data frame (see 'Note' and 'Examples').

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select-helpers. See 'Examples' or package-vignette.

max.length
Numeric, indicating the maximum length of variable labels in the output. If variable names are longer than max.length, they will be shortened to the last whole word within the first max.length chars.

weights
Bare name, or name as string, of a variable in x that indicates the vector of weights, which will be applied to weight all observations. Default is NULL, so no weights are used.

show
Character vector, indicating which information (columns) that describe the data should be returned. May be one or more of "type", "label", "n", "NA.prc", "mean", "sd", "se", "md", "trimmed", "range", "iqr", "skew". There are two shortcuts: show = "all" (default) shows all information, show = "short" just shows n, missing percentage, mean and standard deviation.

out
Character vector, indicating whether the results should be printed to console (out = "txt") or as HTML-table in the viewer-pane (out = "viewer") or browser (out = "browser").

encoding
Character vector, indicating the charset encoding used for variable and value labels. Default is "UTF-8". Only used when out is not "txt".

file
Destination file, if the output should be saved as file. Only used when out is not "txt".

Value

A data frame with basic descriptive statistics.

Note

data may also be a grouped data frame (see group_by) with up to two grouping variables. Descriptive tables are created for each subgroup then.

Examples

data(efc)
descr(efc, e17age, c160age)
#> Error: Problem with `summarise()` column `var_se`.
#> ℹ `var_se = sqrt(stats::var(rlang::.data$val, na.rm = TRUE)/length(stats::na.omit(rlang::.data$val)))`.
#> ✖ 'x' is NULL

efc$weights <- abs(rnorm(nrow(efc), 1, .3))
descr(efc, c12hour, barthtot, weights = weights)
#> 
#> ## Basic descriptive statistics
#> 
#> 
#> Error in Math.data.frame(structure(list(NA.prc = c(NA_real_, NA_real_, NA_real_), mean = c(NA_real_, NA_real_, NA_real_), sd = c(NA_real_, NA_real_, NA_real_), se = c(NA_real_, NA_real_, NA_real_), skew = c(NA, NA, NA)), class = c("sjmisc_descr", "data.frame"), row.names = c("NA", "NA.1", "NA.2")), digits = 2) non-numeric variable(s) in data frame: skew

library(dplyr)
efc %>% select(e42dep, e15relat, c172code) %>% descr()
#> Error: Problem with `summarise()` column `var_se`.
#> ℹ `var_se = sqrt(stats::var(rlang::.data$val, na.rm = TRUE)/length(stats::na.omit(rlang::.data$val)))`.
#> ✖ 'x' is NULL

# show just a few elements
efc %>% select(e42dep, e15relat, c172code) %>% descr(show = "short")
#> Error: Problem with `summarise()` column `var_se`.
#> ℹ `var_se = sqrt(stats::var(rlang::.data$val, na.rm = TRUE)/length(stats::na.omit(rlang::.data$val)))`.
#> ✖ 'x' is NULL

# with grouped data frames
efc %>%
  group_by(e16sex) %>%
  select(e16sex, e42dep, e15relat, c172code) %>%
  descr()
#> Error: Problem with `summarise()` column `var_se`.
#> ℹ `var_se = sqrt(stats::var(rlang::.data$val, na.rm = TRUE)/length(stats::na.omit(rlang::.data$val)))`.
#> ✖ 'x' is NULL

# you can select variables also inside 'descr()'
efc %>%
  group_by(e16sex, c172code) %>%
  descr(e16sex, c172code, e17age, c160age)
#> New names:
#> * c172code -> c172code...1
#> * e16sex -> e16sex...2
#> * e16sex -> e16sex...3
#> * c172code -> c172code...4
#> * e16sex -> e16sex...5
#> * ...
#> Error: Problem with `summarise()` column `var_se`.
#> ℹ `var_se = sqrt(stats::var(rlang::.data$val, na.rm = TRUE)/length(stats::na.omit(rlang::.data$val)))`.
#> ✖ 'x' is NULL

# or even use select-helpers
descr(efc, contains("cop"), max.length = 20)
#> Error: Problem with `summarise()` column `var_se`.
#> ℹ `var_se = sqrt(stats::var(rlang::.data$val, na.rm = TRUE)/length(stats::na.omit(rlang::.data$val)))`.
#> ✖ 'x' is NULL