Find variable by name or label

This functions finds variables in a data frame, which variable names or variable (and value) label attribute match a specific pattern. Regular expression for the pattern is supported.

find_var(data, pattern, ignore.case = TRUE, search = c("name_label",
  "name_value", "label_value", "name", "label", "value", "all"),
  out = c("table", "df", "index"), fuzzy = FALSE, as.df, as.varlab)

Arguments

data
A data frame.

pattern
Character string to be matched in data. May also be a character vector of length > 1 (see 'Examples'). pattern is searched for in column names and variable label attributes of data (see get_label). pattern might also be a regular-expression object, as returned by regex, or any of stringr's supported modifiers.

ignore.case
Logical, whether matching should be case sensitive or not.

search
Character string, indicating where pattern is sought. Use one of following options:

"name_label"

The default, searches for pattern in variable names and variable labels.

"name_value"

Searches for pattern in variable names and value labels.

"label_value"

Searches for pattern in variable and value labels.

"name"

Searches for pattern in variable names.

"label"

Searches for pattern in variable labels

"value"

Searches for pattern in value labels.

"all"

Searches for pattern in variable names, variable and value labels.

out
Output (return) format of the search results. May be abbreviated and must be one of:

"table"

A tabular overview (as data frame) with column indices, variable names and labels of matching variables.

"df"

A data frame with all matching variables.

"index"

A named vector with column indices of all matching variables.

fuzzy
Logical, if TRUE, "fuzzy matching" (partial and close distance matching) will be used to find pattern in data if no exact match was found. str_pos is used for fuzzy matching.

as.df
Deprecated, use out = "df" instead.

as.varlab
Deprecated, use out = "table" instead.

Value

By default (i.e. out = "table", returns a tibble with three columns: column number, variable name and variable label. If out = "index", returns a named vector with column indices of matching variables (variable names are used as names-attribute); if out = "df", returns the matching variables as tibble.

Details

This function searches for pattern in data's column names and - for labelled data - in all variable and value labels of data's variables (see get_label for details on variable labels and labelled data). Search is performed using the str_detect functions; hence, regular expressions are supported as well, by simply using pattern = stringr::regex(...).

Examples

data(efc)

# find variables with "cop" in variable name
find_var(efc, "cop")

#> # A tibble: 9 x 3
#>   col.nr var.name var.label                                                    
#>    <int> <chr>    <chr>                                                        
#> 1      6 c82cop1  do you feel you cope well as caregiver?                      
#> 2      7 c83cop2  do you find caregiving too demanding?                        
#> 3      8 c84cop3  does caregiving cause difficulties in your relationship with~
#> 4      9 c85cop4  does caregiving have negative effect on your physical health?
#> 5     10 c86cop5  does caregiving cause difficulties in your relationship with~
#> 6     11 c87cop6  does caregiving cause financial difficulties?                
#> 7     12 c88cop7  do you feel trapped in your role as caregiver?               
#> 8     13 c89cop8  do you feel supported by friends/neighbours?                 
#> 9     14 c90cop9  do you feel caregiving worthwhile?                           

# return tibble with matching variables
find_var(efc, "cop", out = "df")

#> # A tibble: 908 x 9
#>    c82cop1 c83cop2 c84cop3 c85cop4 c86cop5 c87cop6 c88cop7 c89cop8 c90cop9
#>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1       3       2       2       2       1       1       2       3       3
#>  2       3       3       3       3       4       1       3       2       2
#>  3       2       2       1       4       1       1       1       4       3
#>  4       4       1       3       1       1       1       1       2       4
#>  5       3       2       1       2       2       2       1       4       4
#>  6       2       2       3       3       3       2       2       1       1
#>  7       4       2       4       1       1       2       4       1       4
#>  8       3       2       2       1       1       1       2       3       3
#>  9       3       2       3       2       2       1       3       1       3
#> 10       3       2       1       2       1       1       1       1       3
#> # ... with 898 more rows

# or return column numbers
find_var(efc, "cop", out = "index")

#> c82cop1 c83cop2 c84cop3 c85cop4 c86cop5 c87cop6 c88cop7 c89cop8 c90cop9 
#>       6       7       8       9      10      11      12      13      14 

# find variables with "dependency" in names and variable labels
library(sjlabelled)
find_var(efc, "dependency")

#> # A tibble: 1 x 3
#>   col.nr var.name var.label         
#>    <int> <chr>    <chr>             
#> 1      5 e42dep   elder's dependency

get_label(efc$e42dep)

#> [1] "elder's dependency"

# find variables with "level" in names and value labels
res <- find_var(efc, "level", search = "name_value", out = "df")
res

#> # A tibble: 908 x 3
#>    c172code n4pstu nur_pst
#>       <dbl>  <dbl>   <dbl>
#>  1        2      0      NA
#>  2        2      0      NA
#>  3        1      2       2
#>  4        2      3       3
#>  5        2      2       2
#>  6        2      2       2
#>  7        2      3       3
#>  8        2      1       1
#>  9       NA      3       3
#> 10        2      3       3
#> # ... with 898 more rows

get_labels(res, attr.only = FALSE)

#> $c172code
#> [1] "low level of education"          "intermediate level of education"
#> [3] "high level of education"        
#> 
#> $n4pstu
#> [1] "No Care Level" "Care Level 1"  "Care Level 2"  "Care Level 3" 
#> [5] "Care Level 3+"
#> 
#> $nur_pst
#> [1] "Care Level 1"    "Care Level 2"    "Care Level 3/3+"
#> 

# use sjPlot::view_df() to view results

# NOT RUN {
library(sjPlot)
view_df(res)
# }