Frequency table of labelled variables

This function returns a frequency table of labelled vectors, as data frame.

frq(x, ..., sort.frq = c("none", "asc", "desc"), weight.by = NULL,
  auto.grp = NULL, show.strings = TRUE, grp.strings = NULL,
  out = c("txt", "viewer", "browser"))

Arguments

x
A vector or a data frame. May also be a grouped data frame (see 'Note' and 'Examples').

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples' or package-vignette.

sort.frq
Determines whether categories should be sorted according to their frequencies or not. Default is "none", so categories are not sorted by frequency. Use "asc" or "desc" for sorting categories ascending or descending order.

weight.by
Bare name, or name as string, of a variable in x that indicates the vector of weights, which will be applied to weight all observations. Default is NULL, so no weights are used.

auto.grp
Numeric value, indicating the minimum amount of unique values in a variable, at which automatic grouping into smaller units is done (see group_var). Default value for auto.group is NULL, i.e. auto-grouping is off.

show.strings
Logical, if TRUE, frequency tables for character vectors will not be printed. This is useful when printing frequency tables of all variables from a data frame, and due to computational reasons character vectors should not be printed.

grp.strings
Numeric, if not NULL, groups string values in character vectors, based on their similarity. The similarity is estimated with the stringdist-package. See group_str for details on grouping, and that function's maxdist-argument to get more details on the distance of strings to be treated as equal.

out
Character vector, indicating whether the results should be printed to console (out = "txt") or as HTML-table in the viewer-pane (out = "viewer") or browser (out = "browser").

Value

A list of data frames with values, value labels, frequencies, raw, valid and cumulative percentages of x.

Note

x may also be a grouped data frame (see group_by) with up to two grouping variables. Frequency tables are created for each subgroup then.

The print()-method adds a table header with information on the variable label, variable type, total and valid N, and mean and standard deviations. Mean and SD are always printed, even for categorical variables (factors) or character vectors. In this case, values are coerced into numeric vector to calculate the summary statistics.

See also

Examples

library(haven)
# create labelled integer
x <- labelled(
  c(1, 2, 1, 3, 4, 1),
  c(Male = 1, Female = 2, Refused = 3, "N/A" = 4)
)
frq(x)

#> 
#> # x <numeric> 
#> # total N=6  valid N=6  mean=2.00  sd=1.26
#>  
#>  val   label frq raw.prc valid.prc cum.prc
#>    1    Male   3   50.00     50.00   50.00
#>    2  Female   1   16.67     16.67   66.67
#>    3 Refused   1   16.67     16.67   83.33
#>    4     N/A   1   16.67     16.67  100.00
#>   NA      NA   0    0.00        NA      NA
#> 
#> 

x <- labelled(
  c(1:3, tagged_na("a", "c", "z"), 4:1, 2:3),
  c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
    "Refused" = tagged_na("a"), "Not home" = tagged_na("z"))
)
frq(x)

#> 
#> # x <numeric> 
#> # total N=12  valid N=9  mean=2.33  sd=1.00
#>  
#>  val        label frq raw.prc valid.prc cum.prc
#>    1    Agreement   2   16.67     22.22   22.22
#>    2            2   3   25.00     33.33   55.56
#>    3            3   3   25.00     33.33   88.89
#>    4 Disagreement   1    8.33     11.11  100.00
#>   NA           NA   3   25.00        NA      NA
#> 
#> 

# in a pipe
data(efc)
library(dplyr)
efc %>%
  select(e42dep, e15relat, c172code) %>%
  frq()

#> 
#> # elder's dependency (e42dep) <numeric> 
#> # total N=908  valid N=901  mean=2.94  sd=0.94
#>  
#>  val                label frq raw.prc valid.prc cum.prc
#>    1          independent  66    7.27      7.33    7.33
#>    2   slightly dependent 225   24.78     24.97   32.30
#>    3 moderately dependent 306   33.70     33.96   66.26
#>    4   severely dependent 304   33.48     33.74  100.00
#>   NA                   NA   7    0.77        NA      NA
#> 
#> 
#> # relationship to elder (e15relat) <numeric> 
#> # total N=908  valid N=901  mean=2.85  sd=2.08
#>  
#>  val                   label frq raw.prc valid.prc cum.prc
#>    1          spouse/partner 171   18.83     18.98   18.98
#>    2                   child 473   52.09     52.50   71.48
#>    3                 sibling  29    3.19      3.22   74.69
#>    4 daughter or son -in-law  85    9.36      9.43   84.13
#>    5              ancle/aunt  23    2.53      2.55   86.68
#>    6            nephew/niece  22    2.42      2.44   89.12
#>    7                  cousin   6    0.66      0.67   89.79
#>    8          other, specify  92   10.13     10.21  100.00
#>   NA                      NA   7    0.77        NA      NA
#> 
#> 
#> # carer's level of education (c172code) <numeric> 
#> # total N=908  valid N=842  mean=1.97  sd=0.63
#>  
#>  val                           label frq raw.prc valid.prc cum.prc
#>    1          low level of education 180   19.82     21.38   21.38
#>    2 intermediate level of education 506   55.73     60.10   81.47
#>    3         high level of education 156   17.18     18.53  100.00
#>   NA                              NA  66    7.27        NA      NA
#> 
#> 

# or:
# frq(efc, e42dep, e15relat, c172code)

# with grouped data frames, in a pipe
efc %>%
  group_by(e16sex, c172code) %>%
  frq(e16sex, c172code, e42dep)

#> 
#> Grouped by:
#> elder's gender: male
#> carer's level of education: low level of education
#>  
#> # elder's dependency (e42dep) <numeric> 
#> # total N=80  valid N=80  mean=3.06  sd=0.92
#>  
#>  val                label frq raw.prc valid.prc cum.prc
#>    1          independent   5    6.25      6.25    6.25
#>    2   slightly dependent  16   20.00     20.00   26.25
#>    3 moderately dependent  28   35.00     35.00   61.25
#>    4   severely dependent  31   38.75     38.75  100.00
#>   NA                   NA   0    0.00        NA      NA
#> 
#> 
#> Grouped by:
#> elder's gender: male
#> carer's level of education: intermediate level of education
#>  
#> # elder's dependency (e42dep) <numeric> 
#> # total N=156  valid N=156  mean=2.83  sd=0.94
#>  
#>  val                label frq raw.prc valid.prc cum.prc
#>    1          independent  15    9.62      9.62    9.62
#>    2   slightly dependent  39   25.00     25.00   34.62
#>    3 moderately dependent  59   37.82     37.82   72.44
#>    4   severely dependent  43   27.56     27.56  100.00
#>   NA                   NA   0    0.00        NA      NA
#> 
#> 
#> Grouped by:
#> elder's gender: male
#> carer's level of education: high level of education
#>  
#> # elder's dependency (e42dep) <numeric> 
#> # total N=43  valid N=43  mean=2.91  sd=0.81
#>  
#>  val                label frq raw.prc valid.prc cum.prc
#>    1          independent   1    2.33      2.33    2.33
#>    2   slightly dependent  13   30.23     30.23   32.56
#>    3 moderately dependent  18   41.86     41.86   74.42
#>    4   severely dependent  11   25.58     25.58  100.00
#>   NA                   NA   0    0.00        NA      NA
#> 
#> 
#> Grouped by:
#> elder's gender: female
#> carer's level of education: low level of education
#>  
#> # elder's dependency (e42dep) <numeric> 
#> # total N=99  valid N=99  mean=2.95  sd=0.94
#>  
#>  val                label frq raw.prc valid.prc cum.prc
#>    1          independent   7    7.07      7.07    7.07
#>    2   slightly dependent  25   25.25     25.25   32.32
#>    3 moderately dependent  33   33.33     33.33   65.66
#>    4   severely dependent  34   34.34     34.34  100.00
#>   NA                   NA   0    0.00        NA      NA
#> 
#> 
#> Grouped by:
#> elder's gender: female
#> carer's level of education: intermediate level of education
#>  
#> # elder's dependency (e42dep) <numeric> 
#> # total N=350  valid N=350  mean=2.90  sd=0.98
#>  
#>  val                label frq raw.prc valid.prc cum.prc
#>    1          independent  30    8.57      8.57    8.57
#>    2   slightly dependent  96   27.43     27.43   36.00
#>    3 moderately dependent 104   29.71     29.71   65.71
#>    4   severely dependent 120   34.29     34.29  100.00
#>   NA                   NA   0    0.00        NA      NA
#> 
#> 
#> Grouped by:
#> elder's gender: female
#> carer's level of education: high level of education
#>  
#> # elder's dependency (e42dep) <numeric> 
#> # total N=113  valid N=113  mean=3.04  sd=0.85
#>  
#>  val                label frq raw.prc valid.prc cum.prc
#>    1          independent   4    3.54      3.54    3.54
#>    2   slightly dependent  26   23.01     23.01   26.55
#>    3 moderately dependent  44   38.94     38.94   65.49
#>    4   severely dependent  39   34.51     34.51  100.00
#>   NA                   NA   0    0.00        NA      NA
#> 
#> 

# with select-helpers: all variables from the COPE-Index
# (which all have a "cop" in their name)
frq(efc, contains("cop"))

#> 
#> # do you feel you cope well as caregiver? (c82cop1) <numeric> 
#> # total N=908  valid N=901  mean=3.12  sd=0.58
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     never   3    0.33      0.33    0.33
#>    2 sometimes  97   10.68     10.77   11.10
#>    3     often 591   65.09     65.59   76.69
#>    4    always 210   23.13     23.31  100.00
#>   NA        NA   7    0.77        NA      NA
#> 
#> 
#> # do you find caregiving too demanding? (c83cop2) <numeric> 
#> # total N=908  valid N=902  mean=2.02  sd=0.72
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     Never 186   20.48     20.62   20.62
#>    2 Sometimes 547   60.24     60.64   81.26
#>    3     Often 130   14.32     14.41   95.68
#>    4    Always  39    4.30      4.32  100.00
#>   NA        NA   6    0.66        NA      NA
#> 
#> 
#> # does caregiving cause difficulties in your relationship with your friends? (c84cop3) <numeric> 
#> # total N=908  valid N=902  mean=1.63  sd=0.87
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     Never 516   56.83     57.21   57.21
#>    2 Sometimes 252   27.75     27.94   85.14
#>    3     Often  82    9.03      9.09   94.24
#>    4    Always  52    5.73      5.76  100.00
#>   NA        NA   6    0.66        NA      NA
#> 
#> 
#> # does caregiving have negative effect on your physical health? (c85cop4) <numeric> 
#> # total N=908  valid N=898  mean=1.77  sd=0.87
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     Never 409   45.04     45.55   45.55
#>    2 Sometimes 346   38.11     38.53   84.08
#>    3     Often  85    9.36      9.47   93.54
#>    4    Always  58    6.39      6.46  100.00
#>   NA        NA  10    1.10        NA      NA
#> 
#> 
#> # does caregiving cause difficulties in your relationship with your family? (c86cop5) <numeric> 
#> # total N=908  valid N=902  mean=1.39  sd=0.67
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     Never 626   68.94     69.40   69.40
#>    2 Sometimes 211   23.24     23.39   92.79
#>    3     Often  50    5.51      5.54   98.34
#>    4    Always  15    1.65      1.66  100.00
#>   NA        NA   6    0.66        NA      NA
#> 
#> 
#> # does caregiving cause financial difficulties? (c87cop6) <numeric> 
#> # total N=908  valid N=900  mean=1.29  sd=0.64
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     Never 713   78.52     79.22   79.22
#>    2 Sometimes 131   14.43     14.56   93.78
#>    3     Often  39    4.30      4.33   98.11
#>    4    Always  17    1.87      1.89  100.00
#>   NA        NA   8    0.88        NA      NA
#> 
#> 
#> # do you feel trapped in your role as caregiver? (c88cop7) <numeric> 
#> # total N=908  valid N=900  mean=1.92  sd=0.91
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     Never 336   37.00     37.33   37.33
#>    2 Sometimes 374   41.19     41.56   78.89
#>    3     Often 113   12.44     12.56   91.44
#>    4    Always  77    8.48      8.56  100.00
#>   NA        NA   8    0.88        NA      NA
#> 
#> 
#> # do you feel supported by friends/neighbours? (c89cop8) <numeric> 
#> # total N=908  valid N=901  mean=2.16  sd=1.04
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     never 313   34.47     34.74   34.74
#>    2 sometimes 237   26.10     26.30   61.04
#>    3     often 241   26.54     26.75   87.79
#>    4    always 110   12.11     12.21  100.00
#>   NA        NA   7    0.77        NA      NA
#> 
#> 
#> # do you feel caregiving worthwhile? (c90cop9) <numeric> 
#> # total N=908  valid N=888  mean=2.93  sd=0.96
#>  
#>  val     label frq raw.prc valid.prc cum.prc
#>    1     never  76    8.37      8.56    8.56
#>    2 sometimes 210   23.13     23.65   32.21
#>    3     often 300   33.04     33.78   65.99
#>    4    always 302   33.26     34.01  100.00
#>   NA        NA  20    2.20        NA      NA
#> 
#> 

# all variables from column "c161sex" to column "c175empl"
frq(efc, c161sex:c175empl)

#> 
#> # carer's gender (c161sex) <numeric> 
#> # total N=908  valid N=901  mean=1.76  sd=0.43
#>  
#>  val  label frq raw.prc valid.prc cum.prc
#>    1   Male 215   23.68     23.86   23.86
#>    2 Female 686   75.55     76.14  100.00
#>   NA     NA   7    0.77        NA      NA
#> 
#> 
#> # carer's level of education (c172code) <numeric> 
#> # total N=908  valid N=842  mean=1.97  sd=0.63
#>  
#>  val                           label frq raw.prc valid.prc cum.prc
#>    1          low level of education 180   19.82     21.38   21.38
#>    2 intermediate level of education 506   55.73     60.10   81.47
#>    3         high level of education 156   17.18     18.53  100.00
#>   NA                              NA  66    7.27        NA      NA
#> 
#> 
#> # are you currently employed? (c175empl) <numeric> 
#> # total N=908  valid N=902  mean=0.43  sd=0.49
#>  
#>  val label frq raw.prc valid.prc cum.prc
#>    0    no 518   57.05     57.43   57.43
#>    1   yes 384   42.29     42.57  100.00
#>   NA    NA   6    0.66        NA      NA
#> 
#> 

# for non-labelled data, variable name is printed,
# and "label" column is removed from output
data(iris)
frq(iris, Species)

#> 
#> # Species <categorical> 
#> # total N=150  valid N=150  mean=2.00  sd=0.82
#>  
#>         val frq raw.prc valid.prc cum.prc
#>      setosa  50   33.33     33.33   33.33
#>  versicolor  50   33.33     33.33   66.67
#>   virginica  50   33.33     33.33  100.00
#>        <NA>   0    0.00        NA      NA
#> 
#> 

# group variables with large range
frq(efc, c160age)

#> 
#> # carer' age (c160age) <numeric> 
#> # total N=908  valid N=901  mean=53.46  sd=13.35
#>  
#>   val frq raw.prc valid.prc cum.prc
#>    18   1    0.11      0.11    0.11
#>    19   2    0.22      0.22    0.33
#>    20   2    0.22      0.22    0.55
#>    21   1    0.11      0.11    0.67
#>    22   2    0.22      0.22    0.89
#>    23   4    0.44      0.44    1.33
#>    24   3    0.33      0.33    1.66
#>    25   5    0.55      0.55    2.22
#>    26   3    0.33      0.33    2.55
#>    27   4    0.44      0.44    3.00
#>    28   4    0.44      0.44    3.44
#>    29   8    0.88      0.89    4.33
#>    30   9    0.99      1.00    5.33
#>    31   9    0.99      1.00    6.33
#>    32   5    0.55      0.55    6.88
#>    33  10    1.10      1.11    7.99
#>    34   9    0.99      1.00    8.99
#>    35  17    1.87      1.89   10.88
#>    36   6    0.66      0.67   11.54
#>    37   7    0.77      0.78   12.32
#>    38  13    1.43      1.44   13.76
#>    39  12    1.32      1.33   15.09
#>    40  13    1.43      1.44   16.54
#>    41  15    1.65      1.66   18.20
#>    42  21    2.31      2.33   20.53
#>    43  24    2.64      2.66   23.20
#>    44  25    2.75      2.77   25.97
#>    45  14    1.54      1.55   27.52
#>    46  22    2.42      2.44   29.97
#>    47  27    2.97      3.00   32.96
#>    48  26    2.86      2.89   35.85
#>    49  32    3.52      3.55   39.40
#>    50  19    2.09      2.11   41.51
#>    51  15    1.65      1.66   43.17
#>    52  22    2.42      2.44   45.62
#>    53  27    2.97      3.00   48.61
#>    54  28    3.08      3.11   51.72
#>    55  24    2.64      2.66   54.38
#>    56  27    2.97      3.00   57.38
#>    57  21    2.31      2.33   59.71
#>    58  26    2.86      2.89   62.60
#>    59  20    2.20      2.22   64.82
#>    60  28    3.08      3.11   67.92
#>    61  19    2.09      2.11   70.03
#>    62  28    3.08      3.11   73.14
#>    63  25    2.75      2.77   75.92
#>    64  25    2.75      2.77   78.69
#>    65  26    2.86      2.89   81.58
#>    66  18    1.98      2.00   83.57
#>    67  17    1.87      1.89   85.46
#>    68  14    1.54      1.55   87.01
#>    69  18    1.98      2.00   89.01
#>    70  15    1.65      1.66   90.68
#>    71   4    0.44      0.44   91.12
#>    72  11    1.21      1.22   92.34
#>    73   7    0.77      0.78   93.12
#>    74   7    0.77      0.78   93.90
#>    75  12    1.32      1.33   95.23
#>    76  10    1.10      1.11   96.34
#>    77   7    0.77      0.78   97.11
#>    78   5    0.55      0.55   97.67
#>    79   5    0.55      0.55   98.22
#>    80   5    0.55      0.55   98.78
#>    81   3    0.33      0.33   99.11
#>    82   1    0.11      0.11   99.22
#>    83   4    0.44      0.44   99.67
#>    85   1    0.11      0.11   99.78
#>    87   1    0.11      0.11   99.89
#>    89   1    0.11      0.11  100.00
#>  <NA>   7    0.77        NA      NA
#> 
#> 

frq(efc, c160age, auto.grp = 5)

#> 
#> # carer' age (c160age) <numeric> 
#> # total N=908  valid N=901  mean=53.46  sd=13.35
#>  
#>  val label frq raw.prc valid.prc cum.prc
#>    1 18-32  62    6.83      6.88    6.88
#>    2 33-47 235   25.88     26.08   32.96
#>    3 48-62 362   39.87     40.18   73.14
#>    4 63-77 216   23.79     23.97   97.11
#>    5 78-92  26    2.86      2.89  100.00
#>   NA    NA   7    0.77        NA      NA
#> 
#> 

# and with weights
efc$weights <- abs(rnorm(n = nrow(efc), mean = 1, sd = .5))
frq(efc, c160age, auto.grp = 5, weight.by = weights)

#> 
#> # carer' age (c160age) <numeric> 
#> # total N=905  valid N=905  mean=53.30  sd=13.36
#>  
#>  val label frq raw.prc valid.prc cum.prc
#>    1 18-32  58    6.41      6.41    6.41
#>    2 33-47 247   27.29     27.29   33.70
#>    3 48-62 361   39.89     39.89   73.59
#>    4 63-77 214   23.65     23.65   97.24
#>    5 78-92  25    2.76      2.76  100.00
#>   NA    NA   0    0.00        NA      NA
#> 
#> 

# group string values

# NOT RUN {
dummy <- efc %>% dplyr::select(3)
dummy$words <- sample(
  c("Hello", "Helo", "Hole", "Apple", "Ape",
    "New", "Old", "System", "Systemic"),
  size = nrow(dummy),
  replace = TRUE
)

frq(dummy)
frq(dummy, grp.strings = 2)
# }