Check if variables or cases have missing / infinite values

This functions checks if variables or observations in a data frame have NA, NaN or Inf values.

has_na(x, ..., by = c("col", "row"), out = c("table", "df", "index"))

incomplete_cases(x, ...)

complete_cases(x, ...)

complete_vars(x, ...)

incomplete_vars(x, ...)

Arguments

x
A data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select-helpers. See 'Examples' or package-vignette.

by
Whether to check column- or row-wise for missing and infinite values. If by = "col", has_na() checks for NA/NaN/Inf in columns; If by = "row", has_na() checks each row for these values.

out
Output (return) format of the results. May be abbreviated.

Value

If x is a vector, returns TRUE if x has any missing or infinite values. If x is a data frame, returns TRUE for each variable (if by = "col") or observation (if by = "row") that has any missing or infinite values. If out = "table", results are returned as data frame, with column number, variable name and label, and a logical vector indicating if a variable has missing values or not. However, it's printed in colors, with green rows indicating that a variable has no missings, while red rows indicate the presence of missings or infinite values. If out = "index", a named vector is returned.

Note

complete_cases() and incomplete_cases() are convenient shortcuts for has_na(by = "row", out = "index"), where the first only returns case-id's for all complete cases, and the latter only for non-complete cases.

complete_vars() and incomplete_vars() are convenient shortcuts for has_na(by = "col", out = "index"), and again only return those column-id's for variables which are (in-)complete.

Examples

data(efc)
has_na(efc$e42dep)
#> [1] TRUE
has_na(efc, e42dep, tot_sc_e, c161sex)
#> ## Variables with missing or infinite values (in red)
#> 
#>    Column       Name         Variable Label
#> 
#>         5     e42dep     elder's dependency
#>        24   tot_sc_e   Services for elderly
#>        16    c161sex         carer's gender
#> 
has_na(efc)
#> ## Variables with missing or infinite values (in red)
#> 
#>    Column       Name                           Variable Label
#> 
#>         1    c12hour   average number of hours of care per...
#>         2   e15relat                    relationship to elder
#>         3     e16sex                           elder's gender
#>         4     e17age                               elder' age
#>         5     e42dep                       elder's dependency
#>         6    c82cop1          do you feel you cope well as...
#>         7    c83cop2            do you find caregiving too...
#>         8    c84cop3    does caregiving cause difficulties...
#>         9    c85cop4         does caregiving have negative...
#>        10    c86cop5    does caregiving cause difficulties...
#>        11    c87cop6       does caregiving cause financial...
#>        12    c88cop7   do you feel trapped in your role as...
#>        13    c89cop8              do you feel supported by...
#>        14    c90cop9       do you feel caregiving worthwhile?
#>        15    c160age                               carer' age
#>        16    c161sex                           carer's gender
#>        17   c172code               carer's level of education
#>        18   c175empl              are you currently employed?
#>        19   barthtot                Total score BARTHEL INDEX
#>        20    neg_c_7             Negative impact with 7 items
#>        21    pos_v_4              Positive value with 4 items
#>        22     quol_5                  Quality of life 5 items
#>        23   resttotn                         Job restrictions
#>        24   tot_sc_e                     Services for elderly
#>        25     n4pstu                               Care level
#>        26    nur_pst                               Care level
#> 

has_na(efc, e42dep, tot_sc_e, c161sex, out = "index")
#>   e42dep tot_sc_e  c161sex 
#>     TRUE    FALSE     TRUE 
has_na(efc, out = "df")
#>            tmp
#> c12hour   TRUE
#> e15relat  TRUE
#> e16sex    TRUE
#> e17age    TRUE
#> e42dep    TRUE
#> c82cop1   TRUE
#> c83cop2   TRUE
#> c84cop3   TRUE
#> c85cop4   TRUE
#> c86cop5   TRUE
#> c87cop6   TRUE
#> c88cop7   TRUE
#> c89cop8   TRUE
#> c90cop9   TRUE
#> c160age   TRUE
#> c161sex   TRUE
#> c172code  TRUE
#> c175empl  TRUE
#> barthtot  TRUE
#> neg_c_7   TRUE
#> pos_v_4   TRUE
#> quol_5    TRUE
#> resttotn FALSE
#> tot_sc_e FALSE
#> n4pstu    TRUE
#> nur_pst   TRUE

has_na(efc, by = "row")
#>     case has.na
#> 1      1   TRUE
#> 2      2   TRUE
#> 3      3  FALSE
#> 4      4  FALSE
#> 5      5  FALSE
#> 6      6  FALSE
#> 7      7  FALSE
#> 8      8  FALSE
#> 9      9   TRUE
#> 10    10  FALSE
#> 11    11  FALSE
#> 12    12  FALSE
#> 13    13  FALSE
#> 14    14  FALSE
#> 15    15   TRUE
#> 16    16   TRUE
#> 17    17   TRUE
#> 18    18  FALSE
#> 19    19  FALSE
#> 20    20  FALSE
#> 21    21  FALSE
#> 22    22  FALSE
#> 23    23  FALSE
#> 24    24  FALSE
#> 25    25  FALSE
#> 26    26  FALSE
#> 27    27   TRUE
#> 28    28   TRUE
#> 29    29  FALSE
#> 30    30   TRUE
#> 31    31   TRUE
#> 32    32  FALSE
#> 33    33  FALSE
#> 34    34  FALSE
#> 35    35  FALSE
#> 36    36  FALSE
#> 37    37  FALSE
#> 38    38   TRUE
#> 39    39  FALSE
#> 40    40  FALSE
#> 41    41  FALSE
#> 42    42   TRUE
#> 43    43  FALSE
#> 44    44  FALSE
#> 45    45  FALSE
#> 46    46   TRUE
#> 47    47   TRUE
#> 48    48   TRUE
#> 49    49   TRUE
#> 50    50  FALSE
#> 51    51  FALSE
#> 52    52   TRUE
#> 53    53  FALSE
#> 54    54  FALSE
#> 55    55  FALSE
#> 56    56  FALSE
#> 57    57   TRUE
#> 58    58   TRUE
#> 59    59  FALSE
#> 60    60  FALSE
#> 61    61   TRUE
#> 62    62  FALSE
#> 63    63   TRUE
#> 64    64  FALSE
#> 65    65   TRUE
#> 66    66  FALSE
#> 67    67  FALSE
#> 68    68  FALSE
#> 69    69   TRUE
#> 70    70   TRUE
#> 71    71  FALSE
#> 72    72   TRUE
#> 73    73  FALSE
#> 74    74   TRUE
#> 75    75  FALSE
#> 76    76  FALSE
#> 77    77  FALSE
#> 78    78  FALSE
#> 79    79  FALSE
#> 80    80  FALSE
#> 81    81   TRUE
#> 82    82   TRUE
#> 83    83  FALSE
#> 84    84   TRUE
#> 85    85  FALSE
#> 86    86  FALSE
#> 87    87  FALSE
#> 88    88  FALSE
#> 89    89  FALSE
#> 90    90   TRUE
#> 91    91  FALSE
#> 92    92   TRUE
#> 93    93   TRUE
#> 94    94   TRUE
#> 95    95   TRUE
#> 96    96   TRUE
#> 97    97   TRUE
#> 98    98   TRUE
#> 99    99   TRUE
#> 100  100  FALSE
#> 101  101   TRUE
#> 102  102  FALSE
#> 103  103   TRUE
#> 104  104   TRUE
#> 105  105  FALSE
#> 106  106   TRUE
#> 107  107  FALSE
#> 108  108   TRUE
#> 109  109  FALSE
#> 110  110  FALSE
#> 111  111  FALSE
#> 112  112  FALSE
#> 113  113  FALSE
#> 114  114  FALSE
#> 115  115  FALSE
#> 116  116  FALSE
#> 117  117  FALSE
#> 118  118  FALSE
#> 119  119  FALSE
#> 120  120  FALSE
#> 121  121  FALSE
#> 122  122   TRUE
#> 123  123   TRUE
#> 124  124  FALSE
#> 125  125   TRUE
#> 126  126  FALSE
#> 127  127   TRUE
#> 128  128   TRUE
#> 129  129   TRUE
#> 130  130  FALSE
#> 131  131  FALSE
#> 132  132  FALSE
#> 133  133  FALSE
#> 134  134   TRUE
#> 135  135  FALSE
#> 136  136  FALSE
#> 137  137  FALSE
#> 138  138   TRUE
#> 139  139  FALSE
#> 140  140  FALSE
#> 141  141   TRUE
#> 142  142   TRUE
#> 143  143   TRUE
#> 144  144   TRUE
#> 145  145  FALSE
#> 146  146  FALSE
#> 147  147   TRUE
#> 148  148   TRUE
#> 149  149  FALSE
#> 150  150  FALSE
#> 151  151  FALSE
#> 152  152  FALSE
#> 153  153  FALSE
#> 154  154  FALSE
#> 155  155  FALSE
#> 156  156  FALSE
#> 157  157  FALSE
#> 158  158  FALSE
#> 159  159  FALSE
#> 160  160  FALSE
#> 161  161  FALSE
#> 162  162  FALSE
#> 163  163   TRUE
#> 164  164   TRUE
#> 165  165   TRUE
#> 166  166  FALSE
#> 167  167   TRUE
#> 168  168   TRUE
#> 169  169  FALSE
#> 170  170   TRUE
#> 171  171   TRUE
#> 172  172  FALSE
#> 173  173   TRUE
#> 174  174  FALSE
#> 175  175  FALSE
#> 176  176  FALSE
#> 177  177  FALSE
#> 178  178   TRUE
#> 179  179  FALSE
#> 180  180  FALSE
#> 181  181  FALSE
#> 182  182  FALSE
#> 183  183   TRUE
#> 184  184   TRUE
#> 185  185   TRUE
#> 186  186  FALSE
#> 187  187   TRUE
#> 188  188  FALSE
#> 189  189   TRUE
#> 190  190  FALSE
#> 191  191  FALSE
#> 192  192   TRUE
#> 193  193   TRUE
#> 194  194  FALSE
#> 195  195  FALSE
#> 196  196   TRUE
#> 197  197   TRUE
#> 198  198   TRUE
#> 199  199   TRUE
#> 200  200   TRUE
#> 201  201   TRUE
#> 202  202  FALSE
#> 203  203   TRUE
#> 204  204   TRUE
#> 205  205  FALSE
#> 206  206  FALSE
#> 207  207  FALSE
#> 208  208  FALSE
#> 209  209  FALSE
#> 210  210   TRUE
#> 211  211  FALSE
#> 212  212  FALSE
#> 213  213  FALSE
#> 214  214  FALSE
#> 215  215   TRUE
#> 216  216  FALSE
#> 217  217  FALSE
#> 218  218  FALSE
#> 219  219  FALSE
#> 220  220   TRUE
#> 221  221  FALSE
#> 222  222  FALSE
#> 223  223   TRUE
#> 224  224  FALSE
#> 225  225  FALSE
#> 226  226  FALSE
#> 227  227  FALSE
#> 228  228   TRUE
#> 229  229  FALSE
#> 230  230  FALSE
#> 231  231   TRUE
#> 232  232   TRUE
#> 233  233   TRUE
#> 234  234   TRUE
#> 235  235   TRUE
#> 236  236   TRUE
#> 237  237  FALSE
#> 238  238   TRUE
#> 239  239  FALSE
#> 240  240  FALSE
#> 241  241   TRUE
#> 242  242   TRUE
#> 243  243   TRUE
#> 244  244   TRUE
#> 245  245  FALSE
#> 246  246  FALSE
#> 247  247  FALSE
#> 248  248   TRUE
#> 249  249   TRUE
#> 250  250   TRUE
#> 251  251   TRUE
#> 252  252  FALSE
#> 253  253   TRUE
#> 254  254   TRUE
#> 255  255  FALSE
#> 256  256   TRUE
#> 257  257  FALSE
#> 258  258   TRUE
#> 259  259  FALSE
#> 260  260   TRUE
#> 261  261   TRUE
#> 262  262   TRUE
#> 263  263  FALSE
#> 264  264  FALSE
#> 265  265  FALSE
#> 266  266   TRUE
#> 267  267  FALSE
#> 268  268   TRUE
#> 269  269   TRUE
#> 270  270  FALSE
#> 271  271  FALSE
#> 272  272   TRUE
#> 273  273  FALSE
#> 274  274   TRUE
#> 275  275   TRUE
#> 276  276   TRUE
#> 277  277   TRUE
#> 278  278   TRUE
#> 279  279   TRUE
#> 280  280   TRUE
#> 281  281   TRUE
#> 282  282   TRUE
#> 283  283   TRUE
#> 284  284   TRUE
#> 285  285   TRUE
#> 286  286   TRUE
#> 287  287   TRUE
#> 288  288  FALSE
#> 289  289   TRUE
#> 290  290   TRUE
#> 291  291  FALSE
#> 292  292   TRUE
#> 293  293  FALSE
#> 294  294   TRUE
#> 295  295  FALSE
#> 296  296   TRUE
#> 297  297   TRUE
#> 298  298  FALSE
#> 299  299   TRUE
#> 300  300   TRUE
#> 301  301   TRUE
#> 302  302  FALSE
#> 303  303   TRUE
#> 304  304   TRUE
#> 305  305   TRUE
#> 306  306  FALSE
#> 307  307  FALSE
#> 308  308  FALSE
#> 309  309   TRUE
#> 310  310   TRUE
#> 311  311   TRUE
#> 312  312  FALSE
#> 313  313  FALSE
#> 314  314  FALSE
#> 315  315   TRUE
#> 316  316  FALSE
#> 317  317  FALSE
#> 318  318  FALSE
#> 319  319  FALSE
#> 320  320  FALSE
#> 321  321   TRUE
#> 322  322  FALSE
#> 323  323   TRUE
#> 324  324   TRUE
#> 325  325  FALSE
#> 326  326   TRUE
#> 327  327  FALSE
#> 328  328   TRUE
#> 329  329  FALSE
#> 330  330   TRUE
#> 331  331   TRUE
#> 332  332   TRUE
#> 333  333  FALSE
#> 334  334  FALSE
#> 335  335  FALSE
#> 336  336   TRUE
#> 337  337   TRUE
#> 338  338  FALSE
#> 339  339  FALSE
#> 340  340  FALSE
#> 341  341   TRUE
#> 342  342   TRUE
#> 343  343   TRUE
#> 344  344   TRUE
#> 345  345   TRUE
#> 346  346   TRUE
#> 347  347  FALSE
#> 348  348   TRUE
#> 349  349  FALSE
#> 350  350   TRUE
#> 351  351   TRUE
#> 352  352  FALSE
#> 353  353   TRUE
#> 354  354   TRUE
#> 355  355   TRUE
#> 356  356  FALSE
#> 357  357  FALSE
#> 358  358   TRUE
#> 359  359   TRUE
#> 360  360   TRUE
#> 361  361   TRUE
#> 362  362  FALSE
#> 363  363   TRUE
#> 364  364   TRUE
#> 365  365  FALSE
#> 366  366   TRUE
#> 367  367   TRUE
#> 368  368   TRUE
#> 369  369   TRUE
#> 370  370   TRUE
#> 371  371   TRUE
#> 372  372   TRUE
#> 373  373   TRUE
#> 374  374  FALSE
#> 375  375   TRUE
#> 376  376   TRUE
#> 377  377  FALSE
#> 378  378  FALSE
#> 379  379   TRUE
#> 380  380  FALSE
#> 381  381  FALSE
#> 382  382  FALSE
#> 383  383  FALSE
#> 384  384  FALSE
#> 385  385   TRUE
#> 386  386   TRUE
#> 387  387   TRUE
#> 388  388   TRUE
#> 389  389  FALSE
#> 390  390   TRUE
#> 391  391  FALSE
#> 392  392  FALSE
#> 393  393   TRUE
#> 394  394   TRUE
#> 395  395  FALSE
#> 396  396  FALSE
#> 397  397  FALSE
#> 398  398  FALSE
#> 399  399  FALSE
#> 400  400  FALSE
#> 401  401  FALSE
#> 402  402  FALSE
#> 403  403  FALSE
#> 404  404  FALSE
#> 405  405   TRUE
#> 406  406   TRUE
#> 407  407   TRUE
#> 408  408   TRUE
#> 409  409   TRUE
#> 410  410  FALSE
#> 411  411  FALSE
#> 412  412  FALSE
#> 413  413   TRUE
#> 414  414   TRUE
#> 415  415   TRUE
#> 416  416   TRUE
#> 417  417   TRUE
#> 418  418   TRUE
#> 419  419  FALSE
#> 420  420  FALSE
#> 421  421   TRUE
#> 422  422  FALSE
#> 423  423  FALSE
#> 424  424  FALSE
#> 425  425   TRUE
#> 426  426  FALSE
#> 427  427  FALSE
#> 428  428  FALSE
#> 429  429  FALSE
#> 430  430  FALSE
#> 431  431  FALSE
#> 432  432   TRUE
#> 433  433  FALSE
#> 434  434   TRUE
#> 435  435   TRUE
#> 436  436   TRUE
#> 437  437   TRUE
#> 438  438  FALSE
#> 439  439   TRUE
#> 440  440  FALSE
#> 441  441   TRUE
#> 442  442   TRUE
#> 443  443   TRUE
#> 444  444   TRUE
#> 445  445   TRUE
#> 446  446   TRUE
#> 447  447  FALSE
#> 448  448   TRUE
#> 449  449  FALSE
#> 450  450  FALSE
#> 451  451  FALSE
#> 452  452   TRUE
#> 453  453   TRUE
#> 454  454   TRUE
#> 455  455  FALSE
#> 456  456  FALSE
#> 457  457  FALSE
#> 458  458   TRUE
#> 459  459  FALSE
#> 460  460   TRUE
#> 461  461   TRUE
#> 462  462  FALSE
#> 463  463  FALSE
#> 464  464  FALSE
#> 465  465  FALSE
#> 466  466  FALSE
#> 467  467  FALSE
#> 468  468  FALSE
#> 469  469  FALSE
#> 470  470   TRUE
#> 471  471   TRUE
#> 472  472  FALSE
#> 473  473   TRUE
#> 474  474   TRUE
#> 475  475   TRUE
#> 476  476   TRUE
#> 477  477   TRUE
#> 478  478  FALSE
#> 479  479   TRUE
#> 480  480   TRUE
#> 481  481  FALSE
#> 482  482  FALSE
#> 483  483  FALSE
#> 484  484  FALSE
#> 485  485  FALSE
#> 486  486  FALSE
#> 487  487   TRUE
#> 488  488   TRUE
#> 489  489   TRUE
#> 490  490   TRUE
#> 491  491   TRUE
#> 492  492   TRUE
#> 493  493   TRUE
#> 494  494  FALSE
#> 495  495   TRUE
#> 496  496   TRUE
#> 497  497   TRUE
#> 498  498   TRUE
#> 499  499   TRUE
#> 500  500  FALSE
#> 501  501   TRUE
#> 502  502   TRUE
#> 503  503  FALSE
#> 504  504   TRUE
#> 505  505   TRUE
#> 506  506  FALSE
#> 507  507  FALSE
#> 508  508  FALSE
#> 509  509  FALSE
#> 510  510  FALSE
#> 511  511  FALSE
#> 512  512   TRUE
#> 513  513  FALSE
#> 514  514   TRUE
#> 515  515   TRUE
#> 516  516   TRUE
#> 517  517   TRUE
#> 518  518   TRUE
#> 519  519  FALSE
#> 520  520  FALSE
#> 521  521   TRUE
#> 522  522   TRUE
#> 523  523  FALSE
#> 524  524  FALSE
#> 525  525   TRUE
#> 526  526   TRUE
#> 527  527  FALSE
#> 528  528   TRUE
#> 529  529   TRUE
#> 530  530   TRUE
#> 531  531   TRUE
#> 532  532   TRUE
#> 533  533   TRUE
#> 534  534  FALSE
#> 535  535  FALSE
#> 536  536   TRUE
#> 537  537   TRUE
#> 538  538   TRUE
#> 539  539  FALSE
#> 540  540   TRUE
#> 541  541   TRUE
#> 542  542  FALSE
#> 543  543  FALSE
#> 544  544  FALSE
#> 545  545   TRUE
#> 546  546   TRUE
#> 547  547   TRUE
#> 548  548   TRUE
#> 549  549  FALSE
#> 550  550  FALSE
#> 551  551   TRUE
#> 552  552  FALSE
#> 553  553   TRUE
#> 554  554   TRUE
#> 555  555   TRUE
#> 556  556   TRUE
#> 557  557   TRUE
#> 558  558  FALSE
#> 559  559  FALSE
#> 560  560   TRUE
#> 561  561   TRUE
#> 562  562   TRUE
#> 563  563   TRUE
#> 564  564  FALSE
#> 565  565  FALSE
#> 566  566  FALSE
#> 567  567   TRUE
#> 568  568   TRUE
#> 569  569   TRUE
#> 570  570  FALSE
#> 571  571  FALSE
#> 572  572  FALSE
#> 573  573  FALSE
#> 574  574  FALSE
#> 575  575   TRUE
#> 576  576  FALSE
#> 577  577   TRUE
#> 578  578  FALSE
#> 579  579   TRUE
#> 580  580  FALSE
#> 581  581   TRUE
#> 582  582  FALSE
#> 583  583   TRUE
#> 584  584  FALSE
#> 585  585  FALSE
#> 586  586   TRUE
#> 587  587   TRUE
#> 588  588   TRUE
#> 589  589  FALSE
#> 590  590  FALSE
#> 591  591  FALSE
#> 592  592   TRUE
#> 593  593  FALSE
#> 594  594  FALSE
#> 595  595  FALSE
#> 596  596  FALSE
#> 597  597   TRUE
#> 598  598  FALSE
#> 599  599  FALSE
#> 600  600  FALSE
#> 601  601  FALSE
#> 602  602  FALSE
#> 603  603  FALSE
#> 604  604  FALSE
#> 605  605   TRUE
#> 606  606   TRUE
#> 607  607   TRUE
#> 608  608   TRUE
#> 609  609   TRUE
#> 610  610   TRUE
#> 611  611  FALSE
#> 612  612  FALSE
#> 613  613  FALSE
#> 614  614  FALSE
#> 615  615   TRUE
#> 616  616   TRUE
#> 617  617  FALSE
#> 618  618   TRUE
#> 619  619  FALSE
#> 620  620  FALSE
#> 621  621  FALSE
#> 622  622   TRUE
#> 623  623   TRUE
#> 624  624  FALSE
#> 625  625   TRUE
#> 626  626   TRUE
#> 627  627   TRUE
#> 628  628  FALSE
#> 629  629  FALSE
#> 630  630   TRUE
#> 631  631   TRUE
#> 632  632   TRUE
#> 633  633  FALSE
#> 634  634   TRUE
#> 635  635  FALSE
#> 636  636   TRUE
#> 637  637   TRUE
#> 638  638   TRUE
#> 639  639   TRUE
#> 640  640   TRUE
#> 641  641   TRUE
#> 642  642  FALSE
#> 643  643  FALSE
#> 644  644  FALSE
#> 645  645   TRUE
#> 646  646   TRUE
#> 647  647   TRUE
#> 648  648   TRUE
#> 649  649   TRUE
#> 650  650  FALSE
#> 651  651  FALSE
#> 652  652   TRUE
#> 653  653  FALSE
#> 654  654   TRUE
#> 655  655  FALSE
#> 656  656  FALSE
#> 657  657  FALSE
#> 658  658   TRUE
#> 659  659   TRUE
#> 660  660   TRUE
#> 661  661   TRUE
#> 662  662   TRUE
#> 663  663   TRUE
#> 664  664   TRUE
#> 665  665   TRUE
#> 666  666   TRUE
#> 667  667   TRUE
#> 668  668  FALSE
#> 669  669  FALSE
#> 670  670   TRUE
#> 671  671   TRUE
#> 672  672   TRUE
#> 673  673   TRUE
#> 674  674   TRUE
#> 675  675  FALSE
#> 676  676   TRUE
#> 677  677   TRUE
#> 678  678  FALSE
#> 679  679  FALSE
#> 680  680   TRUE
#> 681  681  FALSE
#> 682  682  FALSE
#> 683  683   TRUE
#> 684  684  FALSE
#> 685  685   TRUE
#> 686  686   TRUE
#> 687  687   TRUE
#> 688  688  FALSE
#> 689  689   TRUE
#> 690  690   TRUE
#> 691  691   TRUE
#> 692  692   TRUE
#> 693  693   TRUE
#> 694  694  FALSE
#> 695  695  FALSE
#> 696  696   TRUE
#> 697  697  FALSE
#> 698  698  FALSE
#> 699  699  FALSE
#> 700  700  FALSE
#> 701  701  FALSE
#> 702  702  FALSE
#> 703  703  FALSE
#> 704  704  FALSE
#> 705  705   TRUE
#> 706  706   TRUE
#> 707  707   TRUE
#> 708  708   TRUE
#> 709  709   TRUE
#> 710  710  FALSE
#> 711  711   TRUE
#> 712  712   TRUE
#> 713  713  FALSE
#> 714  714   TRUE
#> 715  715   TRUE
#> 716  716   TRUE
#> 717  717   TRUE
#> 718  718  FALSE
#> 719  719   TRUE
#> 720  720   TRUE
#> 721  721  FALSE
#> 722  722   TRUE
#> 723  723   TRUE
#> 724  724  FALSE
#> 725  725   TRUE
#> 726  726  FALSE
#> 727  727  FALSE
#> 728  728  FALSE
#> 729  729   TRUE
#> 730  730  FALSE
#> 731  731  FALSE
#> 732  732   TRUE
#> 733  733   TRUE
#> 734  734  FALSE
#> 735  735   TRUE
#> 736  736   TRUE
#> 737  737   TRUE
#> 738  738  FALSE
#> 739  739   TRUE
#> 740  740   TRUE
#> 741  741   TRUE
#> 742  742  FALSE
#> 743  743   TRUE
#> 744  744   TRUE
#> 745  745  FALSE
#> 746  746  FALSE
#> 747  747  FALSE
#> 748  748   TRUE
#> 749  749  FALSE
#> 750  750   TRUE
#> 751  751   TRUE
#> 752  752   TRUE
#> 753  753  FALSE
#> 754  754   TRUE
#> 755  755  FALSE
#> 756  756  FALSE
#> 757  757  FALSE
#> 758  758  FALSE
#> 759  759   TRUE
#> 760  760   TRUE
#> 761  761   TRUE
#> 762  762   TRUE
#> 763  763  FALSE
#> 764  764   TRUE
#> 765  765   TRUE
#> 766  766  FALSE
#> 767  767   TRUE
#> 768  768  FALSE
#> 769  769   TRUE
#> 770  770   TRUE
#> 771  771  FALSE
#> 772  772  FALSE
#> 773  773   TRUE
#> 774  774  FALSE
#> 775  775   TRUE
#> 776  776   TRUE
#> 777  777   TRUE
#> 778  778   TRUE
#> 779  779  FALSE
#> 780  780  FALSE
#> 781  781  FALSE
#> 782  782  FALSE
#> 783  783   TRUE
#> 784  784   TRUE
#> 785  785   TRUE
#> 786  786   TRUE
#> 787  787   TRUE
#> 788  788   TRUE
#> 789  789  FALSE
#> 790  790  FALSE
#> 791  791  FALSE
#> 792  792  FALSE
#> 793  793  FALSE
#> 794  794   TRUE
#> 795  795   TRUE
#> 796  796   TRUE
#> 797  797   TRUE
#> 798  798  FALSE
#> 799  799  FALSE
#> 800  800   TRUE
#> 801  801   TRUE
#> 802  802   TRUE
#> 803  803   TRUE
#> 804  804  FALSE
#> 805  805   TRUE
#> 806  806   TRUE
#> 807  807  FALSE
#> 808  808  FALSE
#> 809  809   TRUE
#> 810  810   TRUE
#> 811  811   TRUE
#> 812  812   TRUE
#> 813  813   TRUE
#> 814  814   TRUE
#> 815  815   TRUE
#> 816  816   TRUE
#> 817  817   TRUE
#> 818  818  FALSE
#> 819  819   TRUE
#> 820  820   TRUE
#> 821  821   TRUE
#> 822  822   TRUE
#> 823  823   TRUE
#> 824  824   TRUE
#> 825  825   TRUE
#> 826  826   TRUE
#> 827  827   TRUE
#> 828  828   TRUE
#> 829  829   TRUE
#> 830  830   TRUE
#> 831  831   TRUE
#> 832  832   TRUE
#> 833  833   TRUE
#> 834  834   TRUE
#> 835  835   TRUE
#> 836  836   TRUE
#> 837  837   TRUE
#> 838  838   TRUE
#> 839  839   TRUE
#> 840  840   TRUE
#> 841  841   TRUE
#> 842  842  FALSE
#> 843  843   TRUE
#> 844  844   TRUE
#> 845  845   TRUE
#> 846  846  FALSE
#> 847  847   TRUE
#> 848  848   TRUE
#> 849  849   TRUE
#> 850  850   TRUE
#> 851  851   TRUE
#> 852  852  FALSE
#> 853  853   TRUE
#> 854  854   TRUE
#> 855  855   TRUE
#> 856  856   TRUE
#> 857  857   TRUE
#> 858  858   TRUE
#> 859  859   TRUE
#> 860  860   TRUE
#> 861  861  FALSE
#> 862  862   TRUE
#> 863  863   TRUE
#> 864  864   TRUE
#> 865  865   TRUE
#> 866  866   TRUE
#> 867  867   TRUE
#> 868  868   TRUE
#> 869  869   TRUE
#> 870  870  FALSE
#> 871  871   TRUE
#> 872  872   TRUE
#> 873  873   TRUE
#> 874  874   TRUE
#> 875  875   TRUE
#> 876  876   TRUE
#> 877  877   TRUE
#> 878  878   TRUE
#> 879  879   TRUE
#> 880  880   TRUE
#> 881  881   TRUE
#> 882  882   TRUE
#> 883  883  FALSE
#> 884  884   TRUE
#> 885  885  FALSE
#> 886  886   TRUE
#> 887  887   TRUE
#> 888  888   TRUE
#> 889  889   TRUE
#> 890  890  FALSE
#> 891  891  FALSE
#> 892  892   TRUE
#> 893  893  FALSE
#> 894  894   TRUE
#> 895  895   TRUE
#> 896  896   TRUE
#> 897  897  FALSE
#> 898  898  FALSE
#> 899  899  FALSE
#> 900  900  FALSE
#> 901  901   TRUE
#> 902  902   TRUE
#> 903  903   TRUE
#> 904  904   TRUE
#> 905  905   TRUE
#> 906  906   TRUE
#> 907  907   TRUE
#> 908  908   TRUE
has_na(efc, e42dep, tot_sc_e, c161sex, by = "row", out = "index")
#>  97 279 903 904 905 906 907 908 
#>  97 279 903 904 905 906 907 908 
has_na(efc, by = "row", out = "df")
#>       tmp
#> 1    TRUE
#> 2    TRUE
#> 3   FALSE
#> 4   FALSE
#> 5   FALSE
#> 6   FALSE
#> 7   FALSE
#> 8   FALSE
#> 9    TRUE
#> 10  FALSE
#> 11  FALSE
#> 12  FALSE
#> 13  FALSE
#> 14  FALSE
#> 15   TRUE
#> 16   TRUE
#> 17   TRUE
#> 18  FALSE
#> 19  FALSE
#> 20  FALSE
#> 21  FALSE
#> 22  FALSE
#> 23  FALSE
#> 24  FALSE
#> 25  FALSE
#> 26  FALSE
#> 27   TRUE
#> 28   TRUE
#> 29  FALSE
#> 30   TRUE
#> 31   TRUE
#> 32  FALSE
#> 33  FALSE
#> 34  FALSE
#> 35  FALSE
#> 36  FALSE
#> 37  FALSE
#> 38   TRUE
#> 39  FALSE
#> 40  FALSE
#> 41  FALSE
#> 42   TRUE
#> 43  FALSE
#> 44  FALSE
#> 45  FALSE
#> 46   TRUE
#> 47   TRUE
#> 48   TRUE
#> 49   TRUE
#> 50  FALSE
#> 51  FALSE
#> 52   TRUE
#> 53  FALSE
#> 54  FALSE
#> 55  FALSE
#> 56  FALSE
#> 57   TRUE
#> 58   TRUE
#> 59  FALSE
#> 60  FALSE
#> 61   TRUE
#> 62  FALSE
#> 63   TRUE
#> 64  FALSE
#> 65   TRUE
#> 66  FALSE
#> 67  FALSE
#> 68  FALSE
#> 69   TRUE
#> 70   TRUE
#> 71  FALSE
#> 72   TRUE
#> 73  FALSE
#> 74   TRUE
#> 75  FALSE
#> 76  FALSE
#> 77  FALSE
#> 78  FALSE
#> 79  FALSE
#> 80  FALSE
#> 81   TRUE
#> 82   TRUE
#> 83  FALSE
#> 84   TRUE
#> 85  FALSE
#> 86  FALSE
#> 87  FALSE
#> 88  FALSE
#> 89  FALSE
#> 90   TRUE
#> 91  FALSE
#> 92   TRUE
#> 93   TRUE
#> 94   TRUE
#> 95   TRUE
#> 96   TRUE
#> 97   TRUE
#> 98   TRUE
#> 99   TRUE
#> 100 FALSE
#> 101  TRUE
#> 102 FALSE
#> 103  TRUE
#> 104  TRUE
#> 105 FALSE
#> 106  TRUE
#> 107 FALSE
#> 108  TRUE
#> 109 FALSE
#> 110 FALSE
#> 111 FALSE
#> 112 FALSE
#> 113 FALSE
#> 114 FALSE
#> 115 FALSE
#> 116 FALSE
#> 117 FALSE
#> 118 FALSE
#> 119 FALSE
#> 120 FALSE
#> 121 FALSE
#> 122  TRUE
#> 123  TRUE
#> 124 FALSE
#> 125  TRUE
#> 126 FALSE
#> 127  TRUE
#> 128  TRUE
#> 129  TRUE
#> 130 FALSE
#> 131 FALSE
#> 132 FALSE
#> 133 FALSE
#> 134  TRUE
#> 135 FALSE
#> 136 FALSE
#> 137 FALSE
#> 138  TRUE
#> 139 FALSE
#> 140 FALSE
#> 141  TRUE
#> 142  TRUE
#> 143  TRUE
#> 144  TRUE
#> 145 FALSE
#> 146 FALSE
#> 147  TRUE
#> 148  TRUE
#> 149 FALSE
#> 150 FALSE
#> 151 FALSE
#> 152 FALSE
#> 153 FALSE
#> 154 FALSE
#> 155 FALSE
#> 156 FALSE
#> 157 FALSE
#> 158 FALSE
#> 159 FALSE
#> 160 FALSE
#> 161 FALSE
#> 162 FALSE
#> 163  TRUE
#> 164  TRUE
#> 165  TRUE
#> 166 FALSE
#> 167  TRUE
#> 168  TRUE
#> 169 FALSE
#> 170  TRUE
#> 171  TRUE
#> 172 FALSE
#> 173  TRUE
#> 174 FALSE
#> 175 FALSE
#> 176 FALSE
#> 177 FALSE
#> 178  TRUE
#> 179 FALSE
#> 180 FALSE
#> 181 FALSE
#> 182 FALSE
#> 183  TRUE
#> 184  TRUE
#> 185  TRUE
#> 186 FALSE
#> 187  TRUE
#> 188 FALSE
#> 189  TRUE
#> 190 FALSE
#> 191 FALSE
#> 192  TRUE
#> 193  TRUE
#> 194 FALSE
#> 195 FALSE
#> 196  TRUE
#> 197  TRUE
#> 198  TRUE
#> 199  TRUE
#> 200  TRUE
#> 201  TRUE
#> 202 FALSE
#> 203  TRUE
#> 204  TRUE
#> 205 FALSE
#> 206 FALSE
#> 207 FALSE
#> 208 FALSE
#> 209 FALSE
#> 210  TRUE
#> 211 FALSE
#> 212 FALSE
#> 213 FALSE
#> 214 FALSE
#> 215  TRUE
#> 216 FALSE
#> 217 FALSE
#> 218 FALSE
#> 219 FALSE
#> 220  TRUE
#> 221 FALSE
#> 222 FALSE
#> 223  TRUE
#> 224 FALSE
#> 225 FALSE
#> 226 FALSE
#> 227 FALSE
#> 228  TRUE
#> 229 FALSE
#> 230 FALSE
#> 231  TRUE
#> 232  TRUE
#> 233  TRUE
#> 234  TRUE
#> 235  TRUE
#> 236  TRUE
#> 237 FALSE
#> 238  TRUE
#> 239 FALSE
#> 240 FALSE
#> 241  TRUE
#> 242  TRUE
#> 243  TRUE
#> 244  TRUE
#> 245 FALSE
#> 246 FALSE
#> 247 FALSE
#> 248  TRUE
#> 249  TRUE
#> 250  TRUE
#> 251  TRUE
#> 252 FALSE
#> 253  TRUE
#> 254  TRUE
#> 255 FALSE
#> 256  TRUE
#> 257 FALSE
#> 258  TRUE
#> 259 FALSE
#> 260  TRUE
#> 261  TRUE
#> 262  TRUE
#> 263 FALSE
#> 264 FALSE
#> 265 FALSE
#> 266  TRUE
#> 267 FALSE
#> 268  TRUE
#> 269  TRUE
#> 270 FALSE
#> 271 FALSE
#> 272  TRUE
#> 273 FALSE
#> 274  TRUE
#> 275  TRUE
#> 276  TRUE
#> 277  TRUE
#> 278  TRUE
#> 279  TRUE
#> 280  TRUE
#> 281  TRUE
#> 282  TRUE
#> 283  TRUE
#> 284  TRUE
#> 285  TRUE
#> 286  TRUE
#> 287  TRUE
#> 288 FALSE
#> 289  TRUE
#> 290  TRUE
#> 291 FALSE
#> 292  TRUE
#> 293 FALSE
#> 294  TRUE
#> 295 FALSE
#> 296  TRUE
#> 297  TRUE
#> 298 FALSE
#> 299  TRUE
#> 300  TRUE
#> 301  TRUE
#> 302 FALSE
#> 303  TRUE
#> 304  TRUE
#> 305  TRUE
#> 306 FALSE
#> 307 FALSE
#> 308 FALSE
#> 309  TRUE
#> 310  TRUE
#> 311  TRUE
#> 312 FALSE
#> 313 FALSE
#> 314 FALSE
#> 315  TRUE
#> 316 FALSE
#> 317 FALSE
#> 318 FALSE
#> 319 FALSE
#> 320 FALSE
#> 321  TRUE
#> 322 FALSE
#> 323  TRUE
#> 324  TRUE
#> 325 FALSE
#> 326  TRUE
#> 327 FALSE
#> 328  TRUE
#> 329 FALSE
#> 330  TRUE
#> 331  TRUE
#> 332  TRUE
#> 333 FALSE
#> 334 FALSE
#> 335 FALSE
#> 336  TRUE
#> 337  TRUE
#> 338 FALSE
#> 339 FALSE
#> 340 FALSE
#> 341  TRUE
#> 342  TRUE
#> 343  TRUE
#> 344  TRUE
#> 345  TRUE
#> 346  TRUE
#> 347 FALSE
#> 348  TRUE
#> 349 FALSE
#> 350  TRUE
#> 351  TRUE
#> 352 FALSE
#> 353  TRUE
#> 354  TRUE
#> 355  TRUE
#> 356 FALSE
#> 357 FALSE
#> 358  TRUE
#> 359  TRUE
#> 360  TRUE
#> 361  TRUE
#> 362 FALSE
#> 363  TRUE
#> 364  TRUE
#> 365 FALSE
#> 366  TRUE
#> 367  TRUE
#> 368  TRUE
#> 369  TRUE
#> 370  TRUE
#> 371  TRUE
#> 372  TRUE
#> 373  TRUE
#> 374 FALSE
#> 375  TRUE
#> 376  TRUE
#> 377 FALSE
#> 378 FALSE
#> 379  TRUE
#> 380 FALSE
#> 381 FALSE
#> 382 FALSE
#> 383 FALSE
#> 384 FALSE
#> 385  TRUE
#> 386  TRUE
#> 387  TRUE
#> 388  TRUE
#> 389 FALSE
#> 390  TRUE
#> 391 FALSE
#> 392 FALSE
#> 393  TRUE
#> 394  TRUE
#> 395 FALSE
#> 396 FALSE
#> 397 FALSE
#> 398 FALSE
#> 399 FALSE
#> 400 FALSE
#> 401 FALSE
#> 402 FALSE
#> 403 FALSE
#> 404 FALSE
#> 405  TRUE
#> 406  TRUE
#> 407  TRUE
#> 408  TRUE
#> 409  TRUE
#> 410 FALSE
#> 411 FALSE
#> 412 FALSE
#> 413  TRUE
#> 414  TRUE
#> 415  TRUE
#> 416  TRUE
#> 417  TRUE
#> 418  TRUE
#> 419 FALSE
#> 420 FALSE
#> 421  TRUE
#> 422 FALSE
#> 423 FALSE
#> 424 FALSE
#> 425  TRUE
#> 426 FALSE
#> 427 FALSE
#> 428 FALSE
#> 429 FALSE
#> 430 FALSE
#> 431 FALSE
#> 432  TRUE
#> 433 FALSE
#> 434  TRUE
#> 435  TRUE
#> 436  TRUE
#> 437  TRUE
#> 438 FALSE
#> 439  TRUE
#> 440 FALSE
#> 441  TRUE
#> 442  TRUE
#> 443  TRUE
#> 444  TRUE
#> 445  TRUE
#> 446  TRUE
#> 447 FALSE
#> 448  TRUE
#> 449 FALSE
#> 450 FALSE
#> 451 FALSE
#> 452  TRUE
#> 453  TRUE
#> 454  TRUE
#> 455 FALSE
#> 456 FALSE
#> 457 FALSE
#> 458  TRUE
#> 459 FALSE
#> 460  TRUE
#> 461  TRUE
#> 462 FALSE
#> 463 FALSE
#> 464 FALSE
#> 465 FALSE
#> 466 FALSE
#> 467 FALSE
#> 468 FALSE
#> 469 FALSE
#> 470  TRUE
#> 471  TRUE
#> 472 FALSE
#> 473  TRUE
#> 474  TRUE
#> 475  TRUE
#> 476  TRUE
#> 477  TRUE
#> 478 FALSE
#> 479  TRUE
#> 480  TRUE
#> 481 FALSE
#> 482 FALSE
#> 483 FALSE
#> 484 FALSE
#> 485 FALSE
#> 486 FALSE
#> 487  TRUE
#> 488  TRUE
#> 489  TRUE
#> 490  TRUE
#> 491  TRUE
#> 492  TRUE
#> 493  TRUE
#> 494 FALSE
#> 495  TRUE
#> 496  TRUE
#> 497  TRUE
#> 498  TRUE
#> 499  TRUE
#> 500 FALSE
#> 501  TRUE
#> 502  TRUE
#> 503 FALSE
#> 504  TRUE
#> 505  TRUE
#> 506 FALSE
#> 507 FALSE
#> 508 FALSE
#> 509 FALSE
#> 510 FALSE
#> 511 FALSE
#> 512  TRUE
#> 513 FALSE
#> 514  TRUE
#> 515  TRUE
#> 516  TRUE
#> 517  TRUE
#> 518  TRUE
#> 519 FALSE
#> 520 FALSE
#> 521  TRUE
#> 522  TRUE
#> 523 FALSE
#> 524 FALSE
#> 525  TRUE
#> 526  TRUE
#> 527 FALSE
#> 528  TRUE
#> 529  TRUE
#> 530  TRUE
#> 531  TRUE
#> 532  TRUE
#> 533  TRUE
#> 534 FALSE
#> 535 FALSE
#> 536  TRUE
#> 537  TRUE
#> 538  TRUE
#> 539 FALSE
#> 540  TRUE
#> 541  TRUE
#> 542 FALSE
#> 543 FALSE
#> 544 FALSE
#> 545  TRUE
#> 546  TRUE
#> 547  TRUE
#> 548  TRUE
#> 549 FALSE
#> 550 FALSE
#> 551  TRUE
#> 552 FALSE
#> 553  TRUE
#> 554  TRUE
#> 555  TRUE
#> 556  TRUE
#> 557  TRUE
#> 558 FALSE
#> 559 FALSE
#> 560  TRUE
#> 561  TRUE
#> 562  TRUE
#> 563  TRUE
#> 564 FALSE
#> 565 FALSE
#> 566 FALSE
#> 567  TRUE
#> 568  TRUE
#> 569  TRUE
#> 570 FALSE
#> 571 FALSE
#> 572 FALSE
#> 573 FALSE
#> 574 FALSE
#> 575  TRUE
#> 576 FALSE
#> 577  TRUE
#> 578 FALSE
#> 579  TRUE
#> 580 FALSE
#> 581  TRUE
#> 582 FALSE
#> 583  TRUE
#> 584 FALSE
#> 585 FALSE
#> 586  TRUE
#> 587  TRUE
#> 588  TRUE
#> 589 FALSE
#> 590 FALSE
#> 591 FALSE
#> 592  TRUE
#> 593 FALSE
#> 594 FALSE
#> 595 FALSE
#> 596 FALSE
#> 597  TRUE
#> 598 FALSE
#> 599 FALSE
#> 600 FALSE
#> 601 FALSE
#> 602 FALSE
#> 603 FALSE
#> 604 FALSE
#> 605  TRUE
#> 606  TRUE
#> 607  TRUE
#> 608  TRUE
#> 609  TRUE
#> 610  TRUE
#> 611 FALSE
#> 612 FALSE
#> 613 FALSE
#> 614 FALSE
#> 615  TRUE
#> 616  TRUE
#> 617 FALSE
#> 618  TRUE
#> 619 FALSE
#> 620 FALSE
#> 621 FALSE
#> 622  TRUE
#> 623  TRUE
#> 624 FALSE
#> 625  TRUE
#> 626  TRUE
#> 627  TRUE
#> 628 FALSE
#> 629 FALSE
#> 630  TRUE
#> 631  TRUE
#> 632  TRUE
#> 633 FALSE
#> 634  TRUE
#> 635 FALSE
#> 636  TRUE
#> 637  TRUE
#> 638  TRUE
#> 639  TRUE
#> 640  TRUE
#> 641  TRUE
#> 642 FALSE
#> 643 FALSE
#> 644 FALSE
#> 645  TRUE
#> 646  TRUE
#> 647  TRUE
#> 648  TRUE
#> 649  TRUE
#> 650 FALSE
#> 651 FALSE
#> 652  TRUE
#> 653 FALSE
#> 654  TRUE
#> 655 FALSE
#> 656 FALSE
#> 657 FALSE
#> 658  TRUE
#> 659  TRUE
#> 660  TRUE
#> 661  TRUE
#> 662  TRUE
#> 663  TRUE
#> 664  TRUE
#> 665  TRUE
#> 666  TRUE
#> 667  TRUE
#> 668 FALSE
#> 669 FALSE
#> 670  TRUE
#> 671  TRUE
#> 672  TRUE
#> 673  TRUE
#> 674  TRUE
#> 675 FALSE
#> 676  TRUE
#> 677  TRUE
#> 678 FALSE
#> 679 FALSE
#> 680  TRUE
#> 681 FALSE
#> 682 FALSE
#> 683  TRUE
#> 684 FALSE
#> 685  TRUE
#> 686  TRUE
#> 687  TRUE
#> 688 FALSE
#> 689  TRUE
#> 690  TRUE
#> 691  TRUE
#> 692  TRUE
#> 693  TRUE
#> 694 FALSE
#> 695 FALSE
#> 696  TRUE
#> 697 FALSE
#> 698 FALSE
#> 699 FALSE
#> 700 FALSE
#> 701 FALSE
#> 702 FALSE
#> 703 FALSE
#> 704 FALSE
#> 705  TRUE
#> 706  TRUE
#> 707  TRUE
#> 708  TRUE
#> 709  TRUE
#> 710 FALSE
#> 711  TRUE
#> 712  TRUE
#> 713 FALSE
#> 714  TRUE
#> 715  TRUE
#> 716  TRUE
#> 717  TRUE
#> 718 FALSE
#> 719  TRUE
#> 720  TRUE
#> 721 FALSE
#> 722  TRUE
#> 723  TRUE
#> 724 FALSE
#> 725  TRUE
#> 726 FALSE
#> 727 FALSE
#> 728 FALSE
#> 729  TRUE
#> 730 FALSE
#> 731 FALSE
#> 732  TRUE
#> 733  TRUE
#> 734 FALSE
#> 735  TRUE
#> 736  TRUE
#> 737  TRUE
#> 738 FALSE
#> 739  TRUE
#> 740  TRUE
#> 741  TRUE
#> 742 FALSE
#> 743  TRUE
#> 744  TRUE
#> 745 FALSE
#> 746 FALSE
#> 747 FALSE
#> 748  TRUE
#> 749 FALSE
#> 750  TRUE
#> 751  TRUE
#> 752  TRUE
#> 753 FALSE
#> 754  TRUE
#> 755 FALSE
#> 756 FALSE
#> 757 FALSE
#> 758 FALSE
#> 759  TRUE
#> 760  TRUE
#> 761  TRUE
#> 762  TRUE
#> 763 FALSE
#> 764  TRUE
#> 765  TRUE
#> 766 FALSE
#> 767  TRUE
#> 768 FALSE
#> 769  TRUE
#> 770  TRUE
#> 771 FALSE
#> 772 FALSE
#> 773  TRUE
#> 774 FALSE
#> 775  TRUE
#> 776  TRUE
#> 777  TRUE
#> 778  TRUE
#> 779 FALSE
#> 780 FALSE
#> 781 FALSE
#> 782 FALSE
#> 783  TRUE
#> 784  TRUE
#> 785  TRUE
#> 786  TRUE
#> 787  TRUE
#> 788  TRUE
#> 789 FALSE
#> 790 FALSE
#> 791 FALSE
#> 792 FALSE
#> 793 FALSE
#> 794  TRUE
#> 795  TRUE
#> 796  TRUE
#> 797  TRUE
#> 798 FALSE
#> 799 FALSE
#> 800  TRUE
#> 801  TRUE
#> 802  TRUE
#> 803  TRUE
#> 804 FALSE
#> 805  TRUE
#> 806  TRUE
#> 807 FALSE
#> 808 FALSE
#> 809  TRUE
#> 810  TRUE
#> 811  TRUE
#> 812  TRUE
#> 813  TRUE
#> 814  TRUE
#> 815  TRUE
#> 816  TRUE
#> 817  TRUE
#> 818 FALSE
#> 819  TRUE
#> 820  TRUE
#> 821  TRUE
#> 822  TRUE
#> 823  TRUE
#> 824  TRUE
#> 825  TRUE
#> 826  TRUE
#> 827  TRUE
#> 828  TRUE
#> 829  TRUE
#> 830  TRUE
#> 831  TRUE
#> 832  TRUE
#> 833  TRUE
#> 834  TRUE
#> 835  TRUE
#> 836  TRUE
#> 837  TRUE
#> 838  TRUE
#> 839  TRUE
#> 840  TRUE
#> 841  TRUE
#> 842 FALSE
#> 843  TRUE
#> 844  TRUE
#> 845  TRUE
#> 846 FALSE
#> 847  TRUE
#> 848  TRUE
#> 849  TRUE
#> 850  TRUE
#> 851  TRUE
#> 852 FALSE
#> 853  TRUE
#> 854  TRUE
#> 855  TRUE
#> 856  TRUE
#> 857  TRUE
#> 858  TRUE
#> 859  TRUE
#> 860  TRUE
#> 861 FALSE
#> 862  TRUE
#> 863  TRUE
#> 864  TRUE
#> 865  TRUE
#> 866  TRUE
#> 867  TRUE
#> 868  TRUE
#> 869  TRUE
#> 870 FALSE
#> 871  TRUE
#> 872  TRUE
#> 873  TRUE
#> 874  TRUE
#> 875  TRUE
#> 876  TRUE
#> 877  TRUE
#> 878  TRUE
#> 879  TRUE
#> 880  TRUE
#> 881  TRUE
#> 882  TRUE
#> 883 FALSE
#> 884  TRUE
#> 885 FALSE
#> 886  TRUE
#> 887  TRUE
#> 888  TRUE
#> 889  TRUE
#> 890 FALSE
#> 891 FALSE
#> 892  TRUE
#> 893 FALSE
#> 894  TRUE
#> 895  TRUE
#> 896  TRUE
#> 897 FALSE
#> 898 FALSE
#> 899 FALSE
#> 900 FALSE
#> 901  TRUE
#> 902  TRUE
#> 903  TRUE
#> 904  TRUE
#> 905  TRUE
#> 906  TRUE
#> 907  TRUE
#> 908  TRUE

complete_cases(efc, e42dep, tot_sc_e, c161sex)
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  98  99 100 101 102 103 104 105 106 107 108 109
#> [109] 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126 127
#> [127] 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144 145
#> [145] 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162 163
#> [163] 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180 181
#> [181] 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198 199
#> [199] 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217
#> [217] 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234 235
#> [235] 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252 253
#> [253] 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270 271
#> [271] 272 273 274 275 276 277 278 280 281 282 283 284 285 286 287 288 289 290
#> [289] 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306 307 308
#> [307] 309 310 311 312 313 314 315 316 317 318 319 320 321 322 323 324 325 326
#> [325] 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342 343 344
#> [343] 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359 360 361 362
#> [361] 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379 380
#> [379] 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398
#> [397] 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416
#> [415] 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432 433 434
#> [433] 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452
#> [451] 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468 469 470
#> [469] 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486 487 488
#> [487] 489 490 491 492 493 494 495 496 497 498 499 500 501 502 503 504 505 506
#> [505] 507 508 509 510 511 512 513 514 515 516 517 518 519 520 521 522 523 524
#> [523] 525 526 527 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542
#> [541] 543 544 545 546 547 548 549 550 551 552 553 554 555 556 557 558 559 560
#> [559] 561 562 563 564 565 566 567 568 569 570 571 572 573 574 575 576 577 578
#> [577] 579 580 581 582 583 584 585 586 587 588 589 590 591 592 593 594 595 596
#> [595] 597 598 599 600 601 602 603 604 605 606 607 608 609 610 611 612 613 614
#> [613] 615 616 617 618 619 620 621 622 623 624 625 626 627 628 629 630 631 632
#> [631] 633 634 635 636 637 638 639 640 641 642 643 644 645 646 647 648 649 650
#> [649] 651 652 653 654 655 656 657 658 659 660 661 662 663 664 665 666 667 668
#> [667] 669 670 671 672 673 674 675 676 677 678 679 680 681 682 683 684 685 686
#> [685] 687 688 689 690 691 692 693 694 695 696 697 698 699 700 701 702 703 704
#> [703] 705 706 707 708 709 710 711 712 713 714 715 716 717 718 719 720 721 722
#> [721] 723 724 725 726 727 728 729 730 731 732 733 734 735 736 737 738 739 740
#> [739] 741 742 743 744 745 746 747 748 749 750 751 752 753 754 755 756 757 758
#> [757] 759 760 761 762 763 764 765 766 767 768 769 770 771 772 773 774 775 776
#> [775] 777 778 779 780 781 782 783 784 785 786 787 788 789 790 791 792 793 794
#> [793] 795 796 797 798 799 800 801 802 803 804 805 806 807 808 809 810 811 812
#> [811] 813 814 815 816 817 818 819 820 821 822 823 824 825 826 827 828 829 830
#> [829] 831 832 833 834 835 836 837 838 839 840 841 842 843 844 845 846 847 848
#> [847] 849 850 851 852 853 854 855 856 857 858 859 860 861 862 863 864 865 866
#> [865] 867 868 869 870 871 872 873 874 875 876 877 878 879 880 881 882 883 884
#> [883] 885 886 887 888 889 890 891 892 893 894 895 896 897 898 899 900 901 902
incomplete_cases(efc, e42dep, tot_sc_e, c161sex)
#>  97 279 903 904 905 906 907 908 
#>  97 279 903 904 905 906 907 908 
complete_vars(efc, e42dep, tot_sc_e, c161sex)
#> tot_sc_e 
#>        2 
incomplete_vars(efc, e42dep, tot_sc_e, c161sex)
#>  e42dep c161sex 
#>       1       3