Recode variable categories into new values

Recodes (or "renumbers") the categories of variables into new category values, beginning with the lowest value specified by lowest. Useful when recoding dummy variables with 1/2 values to 0/1 values, or recoding scales from 1-4 to 0-3 etc. recode_to_if() is a scoped variant of recode_to(), where recoding will be applied only to those variables that match the logical condition of predicate.

Usage,
recode_to(x, ..., lowest = 0, highest = -1, append = TRUE, suffix = "_r0")

recode_to_if(
  x,
  predicate,
  lowest = 0,
  highest = -1,
  append = TRUE,
  suffix = "_r0"
)

Arguments

x

A vector or data frame.

...

Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select-helpers. See 'Examples' or package-vignette.

lowest

Indicating the lowest category value for recoding. Default is 0, so the new variable starts with value 0.

highest

If specified and greater than lowest, all category values larger than highest will be set to NA. Default is -1, i.e. this argument is ignored and no NA's will be produced.

append

Logical, if TRUE (the default) and x is a data frame, x including the new variables as additional columns is returned; if FALSE, only the new variables are returned.

suffix

Indicates which suffix will be added to each dummy variable. Use "numeric" to number dummy variables, e.g. x_1, x_2, x_3 etc. Use "label" to add value label, e.g. x_low, x_mid, x_high. May be abbreviated.

predicate

A predicate function to be applied to the columns. The variables for which predicate returns TRUE are selected.

Value

x with recoded category values, where lowest indicates the lowest value; If x is a data frame, for append = TRUE, x including the recoded variables as new columns is returned; if append = FALSE, only the recoded variables will be returned. If append = TRUE and suffix = "", recoded variables will replace (overwrite) existing variables.

Note

Value and variable label attributes are preserved.

See also

Examples

# recode 1-4 to 0-3
dummy <- sample(1:4, 10, replace = TRUE)
recode_to(dummy)
#>  [1] 3 3 1 2 2 2 2 1 0 3

# recode 3-6 to 0-3
# note that numeric type is returned
dummy <- as.factor(3:6)
recode_to(dummy)
#> [1] 0 1 2 3

# lowest value starting with 1
dummy <- sample(11:15, 10, replace = TRUE)
recode_to(dummy, lowest = 1)
#>  [1] 1 3 4 4 5 2 2 4 2 1

# lowest value starting with 1, highest with 3
# all others set to NA
dummy <- sample(11:15, 10, replace = TRUE)
recode_to(dummy, lowest = 1, highest = 3)
#>  [1] NA  3  1 NA  3 NA  2 NA NA  2

# recode multiple variables at once
data(efc)
recode_to(efc, c82cop1, c83cop2, c84cop3, append = FALSE)
#>     c82cop1_r0 c83cop2_r0 c84cop3_r0
#> 1            2          1          1
#> 2            2          2          2
#> 3            1          1          0
#> 4            3          0          2
#> 5            2          1          0
#> 6            1          1          2
#> 7            3          1          3
#> 8            2          1          1
#> 9            2          1          2
#> 10           2          1          0
#> 11           2          3          3
#> 12           2          2          2
#> 13           2          1          1
#> 14           2          1          3
#> 15           1          2          2
#> 16           3          1          0
#> 17           2          1          0
#> 18           3          1          0
#> 19           2          1          0
#> 20           2          1          3
#> 21           2          1          2
#> 22           3          0          1
#> 23           2          0          0
#> 24           2          1          1
#> 25           2          1          0
#> 26           1          2          3
#> 27           1          2          2
#> 28           1          1          1
#> 29           2          1          1
#> 30           1          1          0
#> 31           2          1          3
#> 32           2          0          2
#> 33           2          1          1
#> 34           1          1          3
#> 35           2          1          0
#> 36           2          1          0
#> 37           2          3          2
#> 38           2          0          0
#> 39           2          1          2
#> 40           1          2          1
#> 41           3          2          1
#> 42           3          1          0
#> 43           2          1          2
#> 44           2          2          1
#> 45           2          3          1
#> 46           1          2          0
#> 47           2          1          2
#> 48           2          2          1
#> 49           2          1          2
#> 50           1          3          3
#> 51           2          2          3
#> 52           3          1          1
#> 53           1          1          3
#> 54           2          1          2
#> 55           3          1          0
#> 56           3          1          1
#> 57          NA          2          3
#> 58           2          2          1
#> 59           1          1          0
#> 60           1          3          3
#> 61           3          1          3
#> 62           2          1          3
#> 63           2          0          0
#> 64           1          1          1
#> 65           2          1          1
#> 66           2          1          1
#> 67           1          2          0
#> 68           2          1          0
#> 69           2          2          0
#> 70           1          2          3
#> 71           3          0          0
#> 72           2          1          1
#> 73           2          1          0
#> 74           1          2          1
#> 75           2          1          1
#> 76           2          1          0
#> 77           1          3          3
#> 78           2          1          0
#> 79           1          2          0
#> 80           2          1          1
#> 81           3          0          0
#> 82           3          1          0
#> 83           2          1          0
#> 84           2          2          1
#> 85           2          1          2
#> 86           2          1          0
#> 87           2          1          0
#> 88           2          2          1
#> 89           2          2          2
#> 90           2          2          1
#> 91           2          0          0
#> 92           2          1          1
#> 93           2          1          0
#> 94           2          2          0
#> 95           2          1          0
#> 96           2          1          0
#> 97           2          1          1
#> 98           2          1          0
#> 99           2          1          1
#> 100          2          2          3
#> 101          2          1          2
#> 102          2          1          1
#> 103          3          1          0
#> 104          2          1          0
#> 105          2          1          1
#> 106          2          1          1
#> 107          3          1          0
#> 108          3          0          0
#> 109          2          1          0
#> 110          2          1          1
#> 111          2          1          0
#> 112          2          1          1
#> 113          2          2          2
#> 114          3          0          0
#> 115          2          1          1
#> 116          2          1          1
#> 117          2          1          1
#> 118          2          2          2
#> 119          3          1          3
#> 120          2          1          0
#> 121          3          0          0
#> 122          2          2          3
#> 123          2          2          3
#> 124          2          3          3
#> 125          1          1          1
#> 126          2          1          0
#> 127          3          0          0
#> 128          2          1          0
#> 129          2          0          0
#> 130          3          0          1
#> 131          1          1          2
#> 132          2          1          0
#> 133          2          1          0
#> 134          0          1          2
#> 135          3          1          1
#> 136          2          1          1
#> 137          2          3          2
#> 138          3          0          0
#> 139          3          0          1
#> 140          2          1          0
#> 141          2          1          1
#> 142          2          1          0
#> 143          1          2          1
#> 144          3          1          0
#> 145          2          1          1
#> 146          2          1          1
#> 147          3          1          0
#> 148          2          1          1
#> 149          2          1          0
#> 150          2          1          1
#> 151          3          1          0
#> 152          3          1          1
#> 153          2          0          0
#> 154          3          1          2
#> 155          3          1          0
#> 156          2          0          1
#> 157          1          0          0
#> 158          3          3          0
#> 159          2          1          1
#> 160          2          1          0
#> 161          2          1          0
#> 162          2          1          3
#> 163          0          2          0
#> 164          2          1          0
#> 165          2          1          0
#> 166          3          2          1
#> 167          3          3          0
#> 168          2          1          0
#> 169          1          2          0
#> 170          2          1          0
#> 171          2          1          0
#> 172          3          1          2
#> 173          2          1          0
#> 174          3          1          1
#> 175          3          0          1
#> 176          2          1          0
#> 177          1          1          0
#> 178          2          1          1
#> 179          3          0          0
#> 180          3          1          0
#> 181          3          1          1
#> 182          2          1          1
#> 183          2          1          2
#> 184          2          1          0
#> 185          2          1          1
#> 186          2          2          0
#> 187          3          2          2
#> 188          1          2          2
#> 189          2          0          0
#> 190          2          1          1
#> 191          2          2          2
#> 192          2          3          0
#> 193          2          1          0
#> 194          2          1          0
#> 195          2          1          0
#> 196          2          0          0
#> 197          2          1          1
#> 198          2          2          0
#> 199          2          1          1
#> 200          2          0          0
#> 201          3          0          0
#> 202          1          2          3
#> 203          3          0          0
#> 204          1          2          1
#> 205          3          0          0
#> 206          2          2          0
#> 207          2          1          0
#> 208          3          1          1
#> 209          2          1          2
#> 210          2          1          0
#> 211          2          1          1
#> 212          1          2          1
#> 213          2          0          1
#> 214          1          1          0
#> 215          3          0          3
#> 216          2          0          0
#> 217          3          1          1
#> 218          2          1          0
#> 219          3          1          1
#> 220          3          0          0
#> 221          2          1          0
#> 222          2          2          2
#> 223          2          1          0
#> 224          2          1          0
#> 225          2          3          0
#> 226          2          1          0
#> 227          2          2          0
#> 228          1          3          1
#> 229          2          0          1
#> 230          2          3          0
#> 231          2          2          2
#> 232          2          3          3
#> 233          3          1          0
#> 234          2          2          0
#> 235          1          1          0
#> 236          2          1          1
#> 237          3          1          1
#> 238          2          1          0
#> 239          3          1          0
#> 240          2          1          0
#> 241          2          0          1
#> 242          2          1          0
#> 243          2          1          1
#> 244          2          1          0
#> 245          2          1          0
#> 246          2          1          0
#> 247          2          0          0
#> 248          2          1          0
#> 249          2          1          0
#> 250          2          1          0
#> 251          2          1          0
#> 252          2          1          1
#> 253          2          1          0
#> 254          2          0          2
#> 255          2          1          1
#> 256          1          2          2
#> 257          2          1          0
#> 258          2          3          0
#> 259          2          1          2
#> 260          2          1          1
#> 261          1          1          0
#> 262          2          1          3
#> 263          2          1          1
#> 264          2          2          1
#> 265          3          0          1
#> 266          2          0          0
#> 267          2          1          3
#> 268          2          3          1
#> 269          2          1          0
#> 270          1          3          2
#> 271          2          0          0
#> 272          2          0          1
#> 273          2          1          1
#> 274          3          0          1
#> 275          2          2          1
#> 276          3          2          0
#> 277          2          1          0
#> 278          2          1          1
#> 279          3          1          0
#> 280          2          1          0
#> 281          2          1          0
#> 282          2          0          0
#> 283          2          1          0
#> 284          2          1          0
#> 285          2          1          0
#> 286          2          0          0
#> 287          1          1          0
#> 288          2          1          0
#> 289          2          2          2
#> 290          1          2          2
#> 291          1          1          0
#> 292          3          0          0
#> 293          3          0          0
#> 294          3          0          0
#> 295          2          1          0
#> 296          2          1          0
#> 297          2          1          2
#> 298          3          0          0
#> 299          2          0          1
#> 300          3          0          0
#> 301          2          1          1
#> 302          1          1          0
#> 303          3          2          1
#> 304          2          1          1
#> 305          2          1          0
#> 306          3          1          3
#> 307          1          1          2
#> 308          2          1          0
#> 309          2          1          0
#> 310          2          0          0
#> 311          2          1          0
#> 312          2          1          1
#> 313          2          1          2
#> 314          2          1          0
#> 315          2          1          0
#> 316          2          2          3
#> 317          2          1          1
#> 318          2          1          1
#> 319          2          1          1
#> 320          3          1          0
#> 321          2          1          0
#> 322          3          2          0
#> 323          2          1          0
#> 324          2          2          0
#> 325          2          1          0
#> 326          2          1          0
#> 327          2          2          0
#> 328          2          1          1
#> 329          3          1          0
#> 330          2          1          0
#> 331          3          1          0
#> 332          2          1          0
#> 333          2          1          1
#> 334          2          0          2
#> 335          3          0          0
#> 336          3          0          1
#> 337          2          2          0
#> 338          3          0          0
#> 339          2          2          0
#> 340          3          0          0
#> 341          2          1          0
#> 342          2          1          0
#> 343          2          1          0
#> 344          3          0          0
#> 345          2          1          0
#> 346          3          0          1
#> 347          2          0          0
#> 348          2          2          3
#> 349          3          0          0
#> 350          3          1          0
#> 351          2          1          0
#> 352          3          0          0
#> 353          2          1          0
#> 354          3          0          1
#> 355          2          0          1
#> 356          1          3          2
#> 357          2          1          0
#> 358          2          1          2
#> 359          2          1          1
#> 360          3          0          0
#> 361          2          1          1
#> 362          2          1          0
#> 363          3          0          0
#> 364          3          1          0
#> 365          3          0          0
#> 366          3          0          0
#> 367          3          0          0
#> 368          2          1          1
#> 369          2          1          1
#> 370          2          0          0
#> 371          2          0          0
#> 372          2          1          0
#> 373          2          1          2
#> 374          1          3          0
#> 375          2          1          1
#> 376          2          2          1
#> 377          3          1          0
#> 378          3          0          0
#> 379          2          0          1
#> 380          2          1          0
#> 381          1          1          3
#> 382          2          1          1
#> 383          1          1          1
#> 384          2          3          1
#> 385          2          1          0
#> 386          2          1          0
#> 387          3          0          0
#> 388          2          1          1
#> 389          2          1          1
#> 390          2          1          1
#> 391          2          2          0
#> 392          1          1          1
#> 393          1          2          0
#> 394          2          1          2
#> 395          2          1          1
#> 396          2          1          2
#> 397          2          1          2
#> 398          2          1          1
#> 399          2          2          0
#> 400          2          0          0
#> 401          2          1          0
#> 402          2          1          0
#> 403          3          3          3
#> 404          3          1          0
#> 405          2          1          0
#> 406          2          1          0
#> 407          2          1          0
#> 408          2          1          0
#> 409          2          1          1
#> 410          2          1          1
#> 411          2          1          1
#> 412          2          1          0
#> 413          2          1          0
#> 414          2          1          0
#> 415          2          1          0
#> 416          1          1          0
#> 417          1          2          0
#> 418          2          1          0
#> 419          2          2          2
#> 420          2          1          0
#> 421          3          1          1
#> 422          2          0          0
#> 423          2          1          1
#> 424          2          1          0
#> 425          2          1          0
#> 426          3          1          3
#> 427          2          1          1
#> 428          2          1          0
#> 429          2          1          1
#> 430          3          1          0
#> 431          2          1          0
#> 432          1          1          1
#> 433          2          1          1
#> 434          2          1          0
#> 435          2          2          0
#> 436          2          1          1
#> 437          1          1          1
#> 438          2          1          1
#> 439          3          0          1
#> 440          2          1          1
#> 441          2          1          1
#> 442          2          1          1
#> 443          2          3          3
#> 444          2          1          0
#> 445          1          1          1
#> 446          2          1          1
#> 447          2          1          1
#> 448          2          1          1
#> 449          2          1          1
#> 450          2          1          1
#> 451          3          1          0
#> 452          2          1          0
#> 453          2          1          0
#> 454          3          0          0
#> 455          3          0          0
#> 456          1          2          1
#> 457          3          0          0
#> 458          2          0          0
#> 459          2          0          0
#> 460          3          1          0
#> 461          2          0          0
#> 462          2          0          0
#> 463          2          1          1
#> 464          2          1          2
#> 465          1          1          0
#> 466          2          3          0
#> 467          2          3          1
#> 468          3          0          0
#> 469          2          0          1
#> 470          3          0          0
#> 471          2          3          0
#> 472          3          0          0
#> 473          2          1          0
#> 474          2          1          0
#> 475          3          0          0
#> 476          3          0          0
#> 477          2          1          0
#> 478          2          2          1
#> 479          3          0          0
#> 480          2          1          0
#> 481          2          2          2
#> 482          2          0          0
#> 483          3          1          0
#> 484          2          2          0
#> 485          2          1          0
#> 486          3          1          0
#> 487          3          1          1
#> 488          2          0          0
#> 489          3          0          0
#> 490          3          1          0
#> 491          3          1          1
#> 492          2          0          1
#> 493          2          1          2
#> 494          2          1          1
#> 495          2          2          1
#> 496          2          0          1
#> 497          2          0          0
#> 498          2          1          1
#> 499          1          3          2
#> 500          3          0          1
#> 501          2          2          2
#> 502          3          1          0
#> 503          3          1          0
#> 504          3          2          2
#> 505          3          1          2
#> 506          2          2          1
#> 507          2          1          1
#> 508          2          2          1
#> 509          2          1          1
#> 510          1          2          1
#> 511          2          1          0
#> 512          1          2          1
#> 513          2          1          1
#> 514          2          0          0
#> 515          3          0          0
#> 516          1          2          2
#> 517          3          0          1
#> 518          2          2          2
#> 519          2          3          1
#> 520          2          1          1
#> 521          2          0          0
#> 522          2          3          3
#> 523          2          1          0
#> 524          3          0          1
#> 525          3          0          0
#> 526          2          0          1
#> 527          2          1          0
#> 528          2          1          0
#> 529          2          0          0
#> 530          2          1          0
#> 531          2          1          1
#> 532          3          0          0
#> 533          2          1          1
#> 534          2          0          0
#> 535          2          1          0
#> 536          3          0          0
#> 537          2          1          0
#> 538          1          1          0
#> 539          1          2          0
#> 540          2          1          1
#> 541          1          1          1
#> 542          2          1          0
#> 543          2          0          1
#> 544          2          0          1
#> 545          3          0          0
#> 546          2          1          0
#> 547          3          0          0
#> 548          2          2          2
#> 549          1          3          2
#> 550          2          0          0
#> 551          3          1          0
#> 552          3          1          0
#> 553          3          1          0
#> 554          3          0          0
#> 555          2          0          0
#> 556          3          0          0
#> 557          2          2          3
#> 558          3          3          3
#> 559          2          1          2
#> 560          3          3          3
#> 561          2          2          1
#> 562          3          1          0
#> 563          2          1          1
#> 564          2          0          0
#> 565          3          1          0
#> 566          3          0          0
#> 567          3          0          0
#> 568          2          1          0
#> 569          2          1          0
#> 570          1          2          3
#> 571          2          2          2
#> 572          2          1          0
#> 573          2          3          1
#> 574          1          1          1
#> 575          3          1          0
#> 576          2          1          0
#> 577          2          1          0
#> 578          2          1          2
#> 579          3          0          0
#> 580          2          1          1
#> 581          2          1          0
#> 582          2          1          1
#> 583          1          1          1
#> 584          1          2          0
#> 585          3          1          1
#> 586          1          2          2
#> 587          2          1          0
#> 588          3          1          1
#> 589          2          1          0
#> 590          2          0          0
#> 591          2          1          0
#> 592          1          2          1
#> 593          2          2          1
#> 594          2          2          0
#> 595          3          0          0
#> 596          2          1          3
#> 597          1          2          2
#> 598          3          1          0
#> 599          1          2          2
#> 600          2          1          1
#> 601          2          1          2
#> 602          2          1          0
#> 603          2          2          3
#> 604          3          0          0
#> 605          2          0          0
#> 606          1          0          0
#> 607          2          1          3
#> 608          2          2          1
#> 609          2          2          2
#> 610          3          0          0
#> 611          2          2          0
#> 612          2          1          1
#> 613          1          1          1
#> 614          3          1          0
#> 615          3          1          3
#> 616          2          1          0
#> 617          3          1          0
#> 618          3          1          0
#> 619          2          1          0
#> 620          2          1          1
#> 621          2          1          0
#> 622          2          0          0
#> 623          2          1          0
#> 624          3          0          0
#> 625          1          2          2
#> 626          2          1          0
#> 627          2          2          0
#> 628          2          2          0
#> 629          2          1          0
#> 630          3          0          0
#> 631          2          0          0
#> 632          2          1          0
#> 633          1          1          0
#> 634          3          1          1
#> 635          2          2          2
#> 636          1          1          1
#> 637          2          1          0
#> 638          2          1          0
#> 639          2          1          0
#> 640          2          1          3
#> 641          2          1          1
#> 642          2          3          1
#> 643          3          1          0
#> 644          3          1          1
#> 645          3          1          0
#> 646          2          1          1
#> 647          2          0          0
#> 648          2          1          0
#> 649          2          1          0
#> 650          2          1          0
#> 651          2          1          0
#> 652          2          1          0
#> 653          2          2          0
#> 654          2          1          0
#> 655          3          0          0
#> 656          3          1          0
#> 657          2          1          1
#> 658          2          1          0
#> 659          3          0          0
#> 660          2          1          2
#> 661          2          0          0
#> 662          2          1          1
#> 663          1          1          1
#> 664          2          1          0
#> 665          2          1          0
#> 666          2          2          0
#> 667          1          2          0
#> 668          2          1          0
#> 669          2          1          0
#> 670          2          1          1
#> 671          2          1          0
#> 672          2          1          0
#> 673          2          1          1
#> 674          2          1          0
#> 675          1          1          3
#> 676          2          1          0
#> 677          2          1          0
#> 678          2          1          1
#> 679          1          1          0
#> 680          2          0          0
#> 681          3          1          0
#> 682          3          0          0
#> 683          2          0          0
#> 684          2          1          0
#> 685          2          1          0
#> 686          2          1          1
#> 687          2          1          0
#> 688          2          1          0
#> 689          1          1          1
#> 690          2          1          1
#> 691          3          0          0
#> 692          2          2          1
#> 693          2          1          1
#> 694          2          1          0
#> 695          3          1          0
#> 696          1          1          1
#> 697          1          3          2
#> 698          2          1          0
#> 699          2          1          0
#> 700          1          2          1
#> 701          2          1          1
#> 702          2          1          1
#> 703          2          1          1
#> 704          2          1          0
#> 705          2          1          0
#> 706          2          1          1
#> 707          2          1          0
#> 708          3          0          0
#> 709          2          1          0
#> 710          3          0          0
#> 711          2          1          0
#> 712          1          3          2
#> 713          2          1          0
#> 714          2          1          1
#> 715          2          1          0
#> 716          2          1          0
#> 717          2          1          1
#> 718          2          1          0
#> 719          2          0          0
#> 720          2          1          1
#> 721          2          1          0
#> 722          2          1          0
#> 723          2          1          0
#> 724          2          1          1
#> 725          2          1          2
#> 726          2          1          1
#> 727          2          1          1
#> 728          2          0          0
#> 729          2          2          1
#> 730          2          1          0
#> 731          2          1          0
#> 732          2          3          2
#> 733          3          1          0
#> 734          2          2          1
#> 735          3          1          0
#> 736          2          1          0
#> 737          2          0          0
#> 738          1          1          0
#> 739          3          1          1
#> 740          2          0          0
#> 741          2          1          2
#> 742          3          0          0
#> 743          1          1          3
#> 744          2          0          0
#> 745          1          1          0
#> 746          3          0          1
#> 747          2          0          0
#> 748          2          1          0
#> 749          2          1          0
#> 750          2          1          1
#> 751          3          0          0
#> 752          3          0          0
#> 753          2          1          1
#> 754          2          1          1
#> 755          1          1          1
#> 756          2          1          1
#> 757          3          1          0
#> 758          2          0          0
#> 759          2          1          0
#> 760          3          1          0
#> 761          2          0          0
#> 762          2          2          1
#> 763          3          0          0
#> 764          2          1          0
#> 765          3          0          0
#> 766          2          1          0
#> 767          2          1          2
#> 768          3          0          0
#> 769          3          0          0
#> 770          3          2          0
#> 771          3          0          0
#> 772          3          0          0
#> 773          2          1          0
#> 774          2          0          0
#> 775          3          0          0
#> 776          1          1          1
#> 777          2          1          0
#> 778          2          1          0
#> 779          3          1          1
#> 780          2          2          0
#> 781          2          1          0
#> 782          1          1          1
#> 783          2          0          0
#> 784          2          0          1
#> 785          2          1          0
#> 786          2          1          0
#> 787          1          2          2
#> 788          2          1          0
#> 789          2          2          0
#> 790          3          1          0
#> 791          3          1          0
#> 792          3          1          0
#> 793          2          0          0
#> 794          2          1          2
#> 795          2          0          0
#> 796          2          1          1
#> 797          2          1          1
#> 798          2          1          0
#> 799          2          0          0
#> 800          2          1          1
#> 801          2          1          0
#> 802          3          2          2
#> 803          2          1          3
#> 804          2          3          0
#> 805          2          1          0
#> 806          2          1          1
#> 807          2          1          0
#> 808          2          1          0
#> 809          3          1          0
#> 810          2          2          0
#> 811          3          1          0
#> 812          2          1          0
#> 813          3          1          0
#> 814          2          1          1
#> 815          3          1          0
#> 816          2          1          0
#> 817          2          2          0
#> 818          2          1          0
#> 819          2          0          0
#> 820          2          1          1
#> 821          3          0          0
#> 822          2          3          2
#> 823          2          2          1
#> 824          1          2          1
#> 825          3          2          1
#> 826          3          1          1
#> 827          2          2          3
#> 828          2          2          0
#> 829          1          1          1
#> 830          2          1          0
#> 831          3          0          0
#> 832          2          2          0
#> 833          3          0          0
#> 834          3          1          0
#> 835          3          1          0
#> 836          2          2          0
#> 837          2          1          0
#> 838          3          0          0
#> 839          2          0          0
#> 840          2          1          3
#> 841          2          1          1
#> 842          2          1          1
#> 843          3          0          1
#> 844          2          0          1
#> 845          3          0          0
#> 846          3          1          0
#> 847          3          0          0
#> 848          2          1          0
#> 849          2          1          0
#> 850          3          1          0
#> 851          2          1          0
#> 852          2          1          0
#> 853          2          2          0
#> 854          3          0          0
#> 855          2          1          0
#> 856          2          1          0
#> 857          3          0          0
#> 858          2          1          1
#> 859          2          1          1
#> 860          3          0          0
#> 861          3          1          0
#> 862          2          1          1
#> 863          2          1          0
#> 864          2          1          0
#> 865          2          1          0
#> 866          2          1          0
#> 867          2          1          0
#> 868          2          1          0
#> 869          2          1          1
#> 870          3          0          0
#> 871          3          0          0
#> 872          2          1          0
#> 873          2          1          0
#> 874          1          1          0
#> 875          3          0          0
#> 876          2          1          1
#> 877          2          1          1
#> 878          2          1          0
#> 879          2          2          3
#> 880          2          1          0
#> 881          2          1          0
#> 882          2          0          0
#> 883          2          1          0
#> 884          1          2          1
#> 885          1          2          1
#> 886          2          2          3
#> 887          0          2          1
#> 888          2          1          0
#> 889          3          1          0
#> 890          2          1          0
#> 891          3          0          0
#> 892          2          1          0
#> 893          3          1          2
#> 894          2          2          3
#> 895          2          1          0
#> 896          2          1          0
#> 897          1          2          0
#> 898          3          0          0
#> 899          2          0          0
#> 900          2          1          0
#> 901          2          0          0
#> 902          2          0          0
#> 903         NA         NA         NA
#> 904         NA         NA         NA
#> 905         NA         NA         NA
#> 906         NA         NA         NA
#> 907         NA         NA         NA
#> 908         NA         NA         NA

library(dplyr)
efc %>%
  select(c82cop1, c83cop2, c84cop3) %>%
  mutate(
    c82new = recode_to(c83cop2, lowest = 5),
    c83new = recode_to(c84cop3, lowest = 3)
  ) %>%
  head()
#>   c82cop1 c83cop2 c84cop3 c82new c83new
#> 1       3       2       2      6      4
#> 2       3       3       3      7      5
#> 3       2       2       1      6      3
#> 4       4       1       3      5      5
#> 5       3       2       1      6      3
#> 6       2       2       3      6      5