Recode variable categories into new values

Recodes (or "renumbers") the categories of variables into new category values, beginning with the lowest value specified by lowest. Useful when recoding dummy variables with 1/2 values to 0/1 values, or recoding scales from 1-4 to 0-3 etc. recode_to_if() is a scoped variant of recode_to(), where recoding will be applied only to those variables that match the logical condition of predicate.

recode_to(x, ..., lowest = 0, highest = -1, append = TRUE,
  suffix = "_r0")

recode_to_if(x, predicate, lowest = 0, highest = -1, append = TRUE,
  suffix = "_r0")

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples' or package-vignette.

lowest
Indicating the lowest category value for recoding. Default is 0, so the new variable starts with value 0.

highest
If specified and greater than lowest, all category values larger than highest will be set to NA. Default is -1, i.e. this argument is ignored and no NA's will be produced.

append
Logical, if TRUE (the default) and x is a data frame, x including the new variables as additional columns is returned; if FALSE, only the new variables are returned.

suffix
String value, will be appended to variable (column) names of x, if x is a data frame. If x is not a data frame, this argument will be ignored. The default value to suffix column names in a data frame depends on the function call:

recoded variables (rec()) will be suffixed with "_r"

recoded variables (recode_to()) will be suffixed with "_r0"

dichotomized variables (dicho()) will be suffixed with "_d"

grouped variables (split_var()) will be suffixed with "_g"

grouped variables (group_var()) will be suffixed with "_gr"

standardized variables (std()) will be suffixed with "_z"

centered variables (center()) will be suffixed with "_c"

predicate
A predicate function to be applied to the columns. The variables for which predicate returns TRUE are selected.

Value

x with recoded category values, where lowest indicates the lowest value; If x is a data frame, for append = TRUE, x including the recoded variables as new columns is returned; if append = FALSE, only the recoded variables will be returned.

Note

Value and variable label attributes are preserved.

See also

Examples

# recode 1-4 to 0-3
dummy <- sample(1:4, 10, replace = TRUE)
recode_to(dummy)

#>  [1] 2 1 1 0 2 1 3 0 1 0

# recode 3-6 to 0-3
# note that numeric type is returned
dummy <- as.factor(3:6)
recode_to(dummy)

#> [1] 0 1 2 3

# lowest value starting with 1
dummy <- sample(11:15, 10, replace = TRUE)
recode_to(dummy, lowest = 1)

#>  [1] 2 3 1 2 2 1 3 5 1 2

# lowest value starting with 1, highest with 3
# all others set to NA
dummy <- sample(11:15, 10, replace = TRUE)
recode_to(dummy, lowest = 1, highest = 3)

#>  [1]  1 NA  1  2  3  3  1  1  3 NA

# recode multiple variables at once
data(efc)
recode_to(efc, c82cop1, c83cop2, c84cop3, append = FALSE)

#> # A tibble: 908 x 3
#>    c82cop1_r0 c83cop2_r0 c84cop3_r0
#>         <dbl>      <dbl>      <dbl>
#>  1          2          1          1
#>  2          2          2          2
#>  3          1          1          0
#>  4          3          0          2
#>  5          2          1          0
#>  6          1          1          2
#>  7          3          1          3
#>  8          2          1          1
#>  9          2          1          2
#> 10          2          1          0
#> # ... with 898 more rows

library(dplyr)
efc %>%
  select(c82cop1, c83cop2, c84cop3) %>%
  mutate(
    c82new = recode_to(c83cop2, lowest = 5),
    c83new = recode_to(c84cop3, lowest = 3)
  ) %>%
  head()

#>   c82cop1 c83cop2 c84cop3 c82new c83new
#> 1       3       2       2      6      4
#> 2       3       3       3      7      5
#> 3       2       2       1      6      3
#> 4       4       1       3      5      5
#> 5       3       2       1      6      3
#> 6       2       2       3      6      5