Remove variables from a data frame

This function removes variables from a data frame, and is intended to use within a pipe-workflow. remove_cols() is an alias for remove_var().

remove_var(x, ...)

remove_cols(x, ...)

Arguments

x
A vector or data frame.

...
Character vector with variable names, or unquoted names of variables that should be removed from the data frame. You may also use functions like : or tidyselect's select_helpers.

Value

x, with variables specified in ... removed.

Examples

mtcars %>% remove_var("disp", "cyl")

#>                      mpg  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4 109 4.11 2.780 18.60  1  1    4    2

mtcars %>% remove_var(c("wt", "vs"))

#>                      mpg cyl  disp  hp drat  qsec am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 16.46  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 17.02  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 18.61  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 19.44  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 17.02  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 20.22  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 15.84  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 20.00  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 22.90  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 18.30  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 18.90  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 17.40  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 17.60  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 18.00  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 17.98  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 17.82  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 17.42  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 19.47  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 18.52  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 19.90  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 20.01  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 16.87  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 17.30  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 15.41  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 17.05  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 18.90  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 16.70  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 16.90  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 14.50  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 15.50  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 14.60  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 18.60  1    4    2

mtcars %>% remove_var(drat:am)

#>                      mpg cyl  disp  hp gear carb
#> Mazda RX4           21.0   6 160.0 110    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110    4    4
#> Datsun 710          22.8   4 108.0  93    4    1
#> Hornet 4 Drive      21.4   6 258.0 110    3    1
#> Hornet Sportabout   18.7   8 360.0 175    3    2
#> Valiant             18.1   6 225.0 105    3    1
#> Duster 360          14.3   8 360.0 245    3    4
#> Merc 240D           24.4   4 146.7  62    4    2
#> Merc 230            22.8   4 140.8  95    4    2
#> Merc 280            19.2   6 167.6 123    4    4
#> Merc 280C           17.8   6 167.6 123    4    4
#> Merc 450SE          16.4   8 275.8 180    3    3
#> Merc 450SL          17.3   8 275.8 180    3    3
#> Merc 450SLC         15.2   8 275.8 180    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205    3    4
#> Lincoln Continental 10.4   8 460.0 215    3    4
#> Chrysler Imperial   14.7   8 440.0 230    3    4
#> Fiat 128            32.4   4  78.7  66    4    1
#> Honda Civic         30.4   4  75.7  52    4    2
#> Toyota Corolla      33.9   4  71.1  65    4    1
#> Toyota Corona       21.5   4 120.1  97    3    1
#> Dodge Challenger    15.5   8 318.0 150    3    2
#> AMC Javelin         15.2   8 304.0 150    3    2
#> Camaro Z28          13.3   8 350.0 245    3    4
#> Pontiac Firebird    19.2   8 400.0 175    3    2
#> Fiat X1-9           27.3   4  79.0  66    4    1
#> Porsche 914-2       26.0   4 120.3  91    5    2
#> Lotus Europa        30.4   4  95.1 113    5    2
#> Ford Pantera L      15.8   8 351.0 264    5    4
#> Ferrari Dino        19.7   6 145.0 175    5    6
#> Maserati Bora       15.0   8 301.0 335    5    8
#> Volvo 142E          21.4   4 121.0 109    4    2