Clean values of character vectors.

This function "cleans" values of a character vector or levels of a factor by removing space and punctuation characters.

tidy_values(x, ...)

clean_values(x, ...)

Arguments

x
A vector or data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples' or package-vignette.

Value

x, with "cleaned" values or levels.

Examples

f1 <- sprintf("Char %s", sample(LETTERS[1:5], size = 10, replace = TRUE))
f2 <- as.factor(sprintf("F / %s", sample(letters[1:5], size = 10, replace = TRUE)))
f3 <- sample(1:5, size = 10, replace = TRUE)

x <- data.frame(f1, f2, f3, stringsAsFactors = FALSE)

clean_values(f1)

#>  [1] "Char_E" "Char_B" "Char_D" "Char_C" "Char_E" "Char_E" "Char_E" "Char_E"
#>  [9] "Char_B" "Char_D"

clean_values(f2)

#>  [1] F_d F_e F_e F_b F_d F_d F_d F_e F_a F_a
#> Levels: F_a F_b F_d F_e

clean_values(x)

#> # A tibble: 10 x 3
#>    f1     f2       f3
#>    <chr>  <fct> <int>
#>  1 Char_E F_d       2
#>  2 Char_B F_e       3
#>  3 Char_D F_e       1
#>  4 Char_C F_b       3
#>  5 Char_E F_d       3
#>  6 Char_E F_d       1
#>  7 Char_E F_d       5
#>  8 Char_E F_e       1
#>  9 Char_B F_a       3
#> 10 Char_D F_a       5