Rename variables

This function renames variables in a data frame, i.e. it renames the columns of the data frame.

var_rename(x, ..., verbose = TRUE)

Arguments

x
A data frame.

...
Pairs of named vectors, where the name (lhs) equals the column name that should be renamed, and the value (rhs) is the new column name.

verbose
Logical, if TRUE, a warning is displayed when variable names do not exist in x.

Value

x, with new column names for those variables specified in ....

Examples

dummy <- data.frame(
  a = sample(1:4, 10, replace = TRUE),
  b = sample(1:4, 10, replace = TRUE),
  c = sample(1:4, 10, replace = TRUE)
)

var_rename(dummy, a = "first.col", c = "3rd.col")

#>    first.col b 3rd.col
#> 1          3 4       3
#> 2          1 3       1
#> 3          3 1       3
#> 4          4 2       2
#> 5          2 1       4
#> 6          1 1       3
#> 7          2 3       4
#> 8          1 1       4
#> 9          1 2       2
#> 10         4 1       2

# using quasi-quotation
library(rlang)

#> 
#> Attaching package: 'rlang'

#> The following objects are masked from 'package:purrr':
#> 
#>     %@%, as_function, flatten, flatten_chr, flatten_dbl, flatten_int,
#>     flatten_lgl, flatten_raw, invoke, list_along, modify, prepend,
#>     splice

#> The following objects are masked from 'package:sjlabelled':
#> 
#>     as_character, as_label

#> The following object is masked from 'package:sjmisc':
#> 
#>     is_empty

#> The following objects are masked from 'package:testthat':
#> 
#>     is_false, is_null, is_true

v1 <- "first.col"
v2 <- "3rd.col"
var_rename(dummy, a = !!v1, c = !!v2)

#>    first.col b 3rd.col
#> 1          3 4       3
#> 2          1 3       1
#> 3          3 1       3
#> 4          4 2       2
#> 5          2 1       4
#> 6          1 1       3
#> 7          2 3       4
#> 8          1 1       4
#> 9          1 2       2
#> 10         4 1       2

x1 <- "a"
x2 <- "b"
var_rename(dummy, !!x1 := !!v1, !!x2 := !!v2)

#>    first.col 3rd.col c
#> 1          3       4 3
#> 2          1       3 1
#> 3          3       1 3
#> 4          4       2 2
#> 5          2       1 4
#> 6          1       1 3
#> 7          2       3 4
#> 8          1       1 4
#> 9          1       2 2
#> 10         4       1 2