Convert infiite or NaN values into regular NA

Replaces all infinite (Inf and -Inf) or NaN values with regular NA.

zap_inf(x, ...)

Arguments

x
A vector or a data frame.

...
Optional, unquoted names of variables that should be selected for further processing. Required, if x is a data frame (and no vector) and only selected variables from x should be processed. You may also use functions like : or tidyselect's select_helpers. See 'Examples' or package-vignette.

Value

x, where all Inf, -Inf and NaN are converted to NA.

Examples

x <- c(1, 2, NA, 3, NaN, 4, NA, 5, Inf, -Inf, 6, 7)
zap_inf(x)

#>  [1]  1  2 NA  3 NA  4 NA  5 NA NA  6  7

data(efc)
# produce some NA and NaN values
efc$e42dep[1] <- NaN
efc$e42dep[2] <- NA
efc$c12hour[1] <- NaN
efc$c12hour[2] <- NA
efc$e17age[2] <- NaN
efc$e17age[1] <- NA

# only zap NaN for c12hour
zap_inf(efc$c12hour)

#>   [1]  NA  NA  70 168 168  16 161 110  28  40 100  25  25  24  56  20  25 126
#>  [19] 168 118 150  50  18 168  15 168  12   7  35 168 150 168 168 119 168 168
#>  [37] 168  28 168  30  14 168 168  50 168  24 168  42 154  60 168 168  24 168
#>  [55] 150 168 168 168 168 168 128 168  11  50  80  15   7  21 168  24   6  30
#>  [73] 168  42  30  85  35  70   9 168  77  24  91   6  22 168 168 168 168  50
#>  [91]  40   9  25  14  49   5   7  21  24   7 168 168   4 168 168  15 168 168
#> [109]  20  59  24   7 100 168 168  28  20  50  89 168  12  24  91 168 168 168
#> [127]  11  35  10 125 140  28  84  15  15  14  24   4 140 168   6  10  20  15
#> [145]  65  35 105  50 168  77 168  42   4 168 168  42   6 168  70  15  30   9
#> [163]  16  35  28 168   9 168  15   6  14  35   7  12  50   6  22  35   6 168
#> [181]  28  10  91   8 168  24 168 168  14  28  16 168   6 168  28 160  39  15
#> [199]  10  10  17 168  70  50  12  10 168 168 168  28  12   8  21  24  15   8
#> [217]  16  56  42   4  25  35  30   6 110  30  24  14   4  20  10   7 161  28
#> [235]  14   4  50  10  18  15   6  17  14   5  10  30  10  10  84  10   6  15
#> [253]  40  21  11   8   5  12  10  20  12  24  27  14  18 168  14  10   6  21
#> [271]  16  15  20  10  22  35  30  16  24 168 168  40  40  50  20  20   6   5
#> [289]  48 168   6   8  43  90 100   8  15   5   8  10  30 168  20  30  30  12
#> [307]  30  30  14  15  50  20  25  40   9  18  25 100   6  25  15  20  10  28
#> [325]  10  30   6  40   7   8  15   4   4 168   7  22  30  15  30  20  25  20
#> [343]  20  25  14  30  20  15   6   5  50   8  20  14  18  16  20  15  25  18
#> [361]  22  18  48  60  10   7   5  35  28   5  25  40  49   4  70   6   7  40
#> [379]  50  30  70  14  70   7   6  14  10  25   5  11 120  18   6  20  15  35
#> [397]  15  15 120  28  10 120 100  18   8  40 140   6  11  15  80   5   6  40
#> [415]   5   7   4  24  28  25 140  14  40   5  20 168  40  80 120 160   7   7
#> [433]   4   6  14   7  28  25  28   7  21  30  25  65  18  40  20  80   7  50
#> [451]  30  10   8   6   6  42  10  30  15  10  10 130 160 168  24  24  30  20
#> [469]  10   7 120   6   6  10  20   5  60  20   6  10  84   8 168  40  40  20
#> [487]  56   6   7  12  10  14  28  12  28  11  16  12  22  17  16  50   6   5
#> [505]  45  16  15  20  20   8   8  15  10  15  27  14  28 168 168  21  12  62
#> [523]  20   7  10  20   8   9  12  10  35   5  18  15   6   8  12  25  10  40
#> [541]  20   6 168   4 168  24  25  40  18  20  40   6   5  30  14   8  28  45
#> [559]  40  30  28  50  84  12  40  10  15  40  24 168 160 162 100 110 140  25
#> [577]  15  35  30  40  10  25  12  35  10  30  10 100  12  56  40  30  60 100
#> [595]  20  35 168  22  60  24  20  70  25   6  18  10  15  20  20  10  22  15
#> [613]  48  80  77   6   6   8  20  20   8   6   8  15  10  10  15   8   6  56
#> [631]   6   8  28  20  70   8   7  10  10  10 168  40  40  30  30  28  14  40
#> [649]  10   4  14  42  42  65   5  10  45  55  50  35  40  15   8  25   5   5
#> [667]   4  14  20  10  50  16  22 100 100  30 120 100  10 120   6   6   5  10
#> [685]   9   9  12  20  10 160 100  20  56   7  35 120  60  10  14  21  10   7
#> [703]  18   4   8  10   6  10  30  10   9  36  15  35  21 168 168  10   6  20
#> [721]   6   8  15  84  45  40  90  35  70   5  20  35   8  30   4  84  10  25
#> [739]  60   6  26  12  25  15  50  18  12  22  21  70 168   5  30  70 168   9
#> [757]   6   4  20   8  14   9   6  14   7   6  20  20  48  10   5   6   8   8
#> [775]   8   8   4   4   8  25  20   6  21  28  70   8  21  28  14   4   5   6
#> [793]   7 140   4   8   7   8   7  30  40  50  35  42  40 168 150 168 168 168
#> [811]  14   6   8  22  10   6  30   8  10  10  12  20  15  20  15  20  25   7
#> [829]   4  20  10  50  20  25   5   7   5  12  35  35  30  40  15  10   8   5
#> [847]  14  14   6  14   7  20   6  10  30  14  10  10  28   8   6  20   5   7
#> [865]  20  18  12  17  14   5   6   4   4   6  10   6  14  30  35  10 100   8
#> [883]  10  20  40  40  40   6  30  28  20  25  30  50  20  15 110  28  85 160
#> [901]  10   8  NA  NA  NA  NA  NA  NA
#> attr(,"label")
#> [1] "average number of hours of care per week"

# only zap NaN for c12hour and e17age, not for e42dep,
# but return complete data framee
zap_inf(efc, c12hour, e17age)

#> # A tibble: 908 x 26
#>    c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4
#>      <dbl>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1      NA        2      2     NA    NaN       3       2       2       2
#>  2      NA        2      2     NA     NA       3       3       3       3
#>  3      70        1      2     82      3       2       2       1       4
#>  4     168        1      2     67      4       4       1       3       1
#>  5     168        2      2     84      4       3       2       1       2
#>  6      16        2      2     85      4       2       2       3       3
#>  7     161        1      1     74      4       4       2       4       1
#>  8     110        4      2     87      4       3       2       2       1
#>  9      28        2      2     79      4       3       2       3       2
#> 10      40        2      2     83      4       3       2       1       2
#> # ... with 898 more rows, and 17 more variables: c86cop5 <dbl>, c87cop6 <dbl>,
#> #   c88cop7 <dbl>, c89cop8 <dbl>, c90cop9 <dbl>, c160age <dbl>, c161sex <dbl>,
#> #   c172code <dbl>, c175empl <dbl>, barthtot <dbl>, neg_c_7 <dbl>,
#> #   pos_v_4 <dbl>, quol_5 <dbl>, resttotn <dbl>, tot_sc_e <dbl>, n4pstu <dbl>,
#> #   nur_pst <dbl>

# zap NaN for complete data frame
zap_inf(efc)

#> # A tibble: 908 x 26
#>    c12hour e15relat e16sex e17age e42dep c82cop1 c83cop2 c84cop3 c85cop4
#>      <dbl>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1      NA        2      2     NA     NA       3       2       2       2
#>  2      NA        2      2     NA     NA       3       3       3       3
#>  3      70        1      2     82      3       2       2       1       4
#>  4     168        1      2     67      4       4       1       3       1
#>  5     168        2      2     84      4       3       2       1       2
#>  6      16        2      2     85      4       2       2       3       3
#>  7     161        1      1     74      4       4       2       4       1
#>  8     110        4      2     87      4       3       2       2       1
#>  9      28        2      2     79      4       3       2       3       2
#> 10      40        2      2     83      4       3       2       1       2
#> # ... with 898 more rows, and 17 more variables: c86cop5 <dbl>, c87cop6 <dbl>,
#> #   c88cop7 <dbl>, c89cop8 <dbl>, c90cop9 <dbl>, c160age <dbl>, c161sex <dbl>,
#> #   c172code <dbl>, c175empl <dbl>, barthtot <dbl>, neg_c_7 <dbl>,
#> #   pos_v_4 <dbl>, quol_5 <dbl>, resttotn <dbl>, tot_sc_e <dbl>, n4pstu <dbl>,
#> #   nur_pst <dbl>